micromegas::analytics::lakehouse

Module partition_cache

In-memory copy of a subnet of the list of the partitions in the db

Structs

LivePartitionProvider

A QueryPartitionProvider that fetches partitions directly from the database.

NullPartitionProvider

A QueryPartitionProvider that always returns an empty list of partitions.

PartitionCache

PartitionCache allows to query partitions based on the insert_time range

Traits

QueryPartitionProvider

A trait for providing queryable partitions.