micromegas::analytics::lakehouse

Module query

property_get function support from SQL Datafusion integration

Functions

make_session_context

query

query_partitions

query_partitions_context

query_partitions_context returns a context to run queries using the partitions as the “source” table

register_extension_functions

register functions that are not depended on the lakehouse architecture

register_functions

register_lakehouse_functions

register functions that are part of the lakehouse architecture