micromegas::perfetto::protos::chrome_legacy_ipc

Enum MessageClass

#[repr(i32)]
pub enum MessageClass {

    ClassUnspecified = 0,
    ClassAutomation = 1,
    ClassFrame = 2,
    ClassPage = 3,
    ClassView = 4,
    ClassWidget = 5,
    ClassInput = 6,
    ClassTest = 7,
    ClassWorker = 8,
    ClassNacl = 9,
    ClassGpuChannel = 10,
    ClassMedia = 11,
    ClassPpapi = 12,
    ClassChrome = 13,
    ClassDrag = 14,
    ClassPrint = 15,
    ClassExtension = 16,
    ClassTextInputClient = 17,
    ClassBlinkTest = 18,
    ClassAccessibility = 19,
    ClassPrerender = 20,
    ClassChromoting = 21,
    ClassBrowserPlugin = 22,
    ClassAndroidWebView = 23,
    ClassNaclHost = 24,
    ClassEncryptedMedia = 25,
    ClassCast = 26,
    ClassGinJavaBridge = 27,
    ClassChromeUtilityPrinting = 28,
    ClassOzoneGpu = 29,
    ClassWebTest = 30,
    ClassNetworkHints = 31,
    ClassExtensionsGuestView = 32,
    ClassGuestView = 33,
    ClassMediaPlayerDelegate = 34,
    ClassExtensionWorker = 35,
    ClassSubresourceFilter = 36,
    ClassUnfreezableFrame = 37,

}

Variants

ClassUnspecified = 0

ClassAutomation = 1

ClassFrame = 2

ClassPage = 3

ClassView = 4

ClassWidget = 5

ClassInput = 6

ClassTest = 7

ClassWorker = 8

ClassNacl = 9

ClassGpuChannel = 10

ClassMedia = 11

ClassPpapi = 12

ClassChrome = 13

ClassDrag = 14

ClassPrint = 15

ClassExtension = 16

ClassTextInputClient = 17

ClassBlinkTest = 18

ClassAccessibility = 19

ClassPrerender = 20

ClassChromoting = 21

ClassBrowserPlugin = 22

ClassAndroidWebView = 23

ClassNaclHost = 24

ClassEncryptedMedia = 25

ClassCast = 26

ClassGinJavaBridge = 27

ClassChromeUtilityPrinting = 28

ClassOzoneGpu = 29

ClassWebTest = 30

ClassNetworkHints = 31

ClassExtensionsGuestView = 32

ClassGuestView = 33

ClassMediaPlayerDelegate = 34

ClassExtensionWorker = 35

ClassSubresourceFilter = 36

ClassUnfreezableFrame = 37

Implementations

impl MessageClass

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of MessageClass.

pub fn from_i32(value: i32) -> Option<MessageClass>

๐Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a MessageClass, or None if value is not a valid variant.

impl MessageClass

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<MessageClass>

Creates an enum from field names used in the ProtoBuf definition.