micromegas::perfetto::protos

Module frame_timeline_event

Nested message and enum types in FrameTimelineEvent.

Structs

ActualDisplayFrameStart

Indicates the start of actual timeline slice for DisplayFrames. Also includes the jank information.

ActualSurfaceFrameStart

Indicates the start of actual timeline slice for SurfaceFrames. Also includes the jank information.

ExpectedDisplayFrameStart

Indicates the start of expected timeline slice for DisplayFrames.

ExpectedSurfaceFrameStart

Indicates the start of expected timeline slice for SurfaceFrames.

FrameEnd

FrameEnd just sends the cookie to indicate that the corresponding <display/surface>frame slice’s end.

Enums

Event

JankSeverityType

Specifies the severity of a jank.

JankType

Specifies the reason(s) most likely to have caused the jank. Used as a bitmask.

PredictionType

Specifies if the predictions for the frame are still valid, expired or unknown.

PresentType

Specifies how a frame was presented on screen w.r.t. timing. Can be different for SurfaceFrame and DisplayFrame.