micromegas::perfetto::protos::ftrace_event

Enum Event

pub enum Event {

    Print(PrintFtraceEvent),
    SchedSwitch(SchedSwitchFtraceEvent),
    CpuFrequency(CpuFrequencyFtraceEvent),
    CpuFrequencyLimits(CpuFrequencyLimitsFtraceEvent),
    CpuIdle(CpuIdleFtraceEvent),
    ClockEnable(ClockEnableFtraceEvent),
    ClockDisable(ClockDisableFtraceEvent),
    ClockSetRate(ClockSetRateFtraceEvent),
    SchedWakeup(SchedWakeupFtraceEvent),
    SchedBlockedReason(SchedBlockedReasonFtraceEvent),
    SchedCpuHotplug(SchedCpuHotplugFtraceEvent),
    SchedWaking(SchedWakingFtraceEvent),
    IpiEntry(IpiEntryFtraceEvent),
    IpiExit(IpiExitFtraceEvent),
    IpiRaise(IpiRaiseFtraceEvent),
    SoftirqEntry(SoftirqEntryFtraceEvent),
    SoftirqExit(SoftirqExitFtraceEvent),
    SoftirqRaise(SoftirqRaiseFtraceEvent),
    I2cRead(I2cReadFtraceEvent),
    I2cWrite(I2cWriteFtraceEvent),
    I2cResult(I2cResultFtraceEvent),
    I2cReply(I2cReplyFtraceEvent),
    SmbusRead(SmbusReadFtraceEvent),
    SmbusWrite(SmbusWriteFtraceEvent),
    SmbusResult(SmbusResultFtraceEvent),
    SmbusReply(SmbusReplyFtraceEvent),
    LowmemoryKill(LowmemoryKillFtraceEvent),
    IrqHandlerEntry(IrqHandlerEntryFtraceEvent),
    IrqHandlerExit(IrqHandlerExitFtraceEvent),
    SyncPt(SyncPtFtraceEvent),
    SyncTimeline(SyncTimelineFtraceEvent),
    SyncWait(SyncWaitFtraceEvent),
    Ext4DaWriteBegin(Ext4DaWriteBeginFtraceEvent),
    Ext4DaWriteEnd(Ext4DaWriteEndFtraceEvent),
    Ext4SyncFileEnter(Ext4SyncFileEnterFtraceEvent),
    Ext4SyncFileExit(Ext4SyncFileExitFtraceEvent),
    BlockRqIssue(BlockRqIssueFtraceEvent),
    MmVmscanDirectReclaimBegin(MmVmscanDirectReclaimBeginFtraceEvent),
    MmVmscanDirectReclaimEnd(MmVmscanDirectReclaimEndFtraceEvent),
    MmVmscanKswapdWake(MmVmscanKswapdWakeFtraceEvent),
    MmVmscanKswapdSleep(MmVmscanKswapdSleepFtraceEvent),
    BinderTransaction(BinderTransactionFtraceEvent),
    BinderTransactionReceived(BinderTransactionReceivedFtraceEvent),
    BinderSetPriority(BinderSetPriorityFtraceEvent),
    BinderLock(BinderLockFtraceEvent),
    BinderLocked(BinderLockedFtraceEvent),
    BinderUnlock(BinderUnlockFtraceEvent),
    WorkqueueActivateWork(WorkqueueActivateWorkFtraceEvent),
    WorkqueueExecuteEnd(WorkqueueExecuteEndFtraceEvent),
    WorkqueueExecuteStart(WorkqueueExecuteStartFtraceEvent),
    WorkqueueQueueWork(WorkqueueQueueWorkFtraceEvent),
    RegulatorDisable(RegulatorDisableFtraceEvent),
    RegulatorDisableComplete(RegulatorDisableCompleteFtraceEvent),
    RegulatorEnable(RegulatorEnableFtraceEvent),
    RegulatorEnableComplete(RegulatorEnableCompleteFtraceEvent),
    RegulatorEnableDelay(RegulatorEnableDelayFtraceEvent),
    RegulatorSetVoltage(RegulatorSetVoltageFtraceEvent),
    RegulatorSetVoltageComplete(RegulatorSetVoltageCompleteFtraceEvent),
    CgroupAttachTask(CgroupAttachTaskFtraceEvent),
    CgroupMkdir(CgroupMkdirFtraceEvent),
    CgroupRemount(CgroupRemountFtraceEvent),
    CgroupRmdir(CgroupRmdirFtraceEvent),
    CgroupTransferTasks(CgroupTransferTasksFtraceEvent),
    CgroupDestroyRoot(CgroupDestroyRootFtraceEvent),
    CgroupRelease(CgroupReleaseFtraceEvent),
    CgroupRename(CgroupRenameFtraceEvent),
    CgroupSetupRoot(CgroupSetupRootFtraceEvent),
    MdpCmdKickoff(MdpCmdKickoffFtraceEvent),
    MdpCommit(MdpCommitFtraceEvent),
    MdpPerfSetOt(MdpPerfSetOtFtraceEvent),
    MdpSsppChange(MdpSsppChangeFtraceEvent),
    TracingMarkWrite(TracingMarkWriteFtraceEvent),
    MdpCmdPingpongDone(MdpCmdPingpongDoneFtraceEvent),
    MdpCompareBw(MdpCompareBwFtraceEvent),
    MdpPerfSetPanicLuts(MdpPerfSetPanicLutsFtraceEvent),
    MdpSsppSet(MdpSsppSetFtraceEvent),
    MdpCmdReadptrDone(MdpCmdReadptrDoneFtraceEvent),
    MdpMisrCrc(MdpMisrCrcFtraceEvent),
    MdpPerfSetQosLuts(MdpPerfSetQosLutsFtraceEvent),
    MdpTraceCounter(MdpTraceCounterFtraceEvent),
    MdpCmdReleaseBw(MdpCmdReleaseBwFtraceEvent),
    MdpMixerUpdate(MdpMixerUpdateFtraceEvent),
    MdpPerfSetWmLevels(MdpPerfSetWmLevelsFtraceEvent),
    MdpVideoUnderrunDone(MdpVideoUnderrunDoneFtraceEvent),
    MdpCmdWaitPingpong(MdpCmdWaitPingpongFtraceEvent),
    MdpPerfPrefillCalc(MdpPerfPrefillCalcFtraceEvent),
    MdpPerfUpdateBus(MdpPerfUpdateBusFtraceEvent),
    RotatorBwAoAsContext(RotatorBwAoAsContextFtraceEvent),
    MmFilemapAddToPageCache(MmFilemapAddToPageCacheFtraceEvent),
    MmFilemapDeleteFromPageCache(MmFilemapDeleteFromPageCacheFtraceEvent),
    MmCompactionBegin(MmCompactionBeginFtraceEvent),
    MmCompactionDeferCompaction(MmCompactionDeferCompactionFtraceEvent),
    MmCompactionDeferred(MmCompactionDeferredFtraceEvent),
    MmCompactionDeferReset(MmCompactionDeferResetFtraceEvent),
    MmCompactionEnd(MmCompactionEndFtraceEvent),
    MmCompactionFinished(MmCompactionFinishedFtraceEvent),
    MmCompactionIsolateFreepages(MmCompactionIsolateFreepagesFtraceEvent),
    MmCompactionIsolateMigratepages(MmCompactionIsolateMigratepagesFtraceEvent),
    MmCompactionKcompactdSleep(MmCompactionKcompactdSleepFtraceEvent),
    MmCompactionKcompactdWake(MmCompactionKcompactdWakeFtraceEvent),
    MmCompactionMigratepages(MmCompactionMigratepagesFtraceEvent),
    MmCompactionSuitable(MmCompactionSuitableFtraceEvent),
    MmCompactionTryToCompactPages(MmCompactionTryToCompactPagesFtraceEvent),
    MmCompactionWakeupKcompactd(MmCompactionWakeupKcompactdFtraceEvent),
    SuspendResume(SuspendResumeFtraceEvent),
    SchedWakeupNew(SchedWakeupNewFtraceEvent),
    BlockBioBackmerge(BlockBioBackmergeFtraceEvent),
    BlockBioBounce(BlockBioBounceFtraceEvent),
    BlockBioComplete(BlockBioCompleteFtraceEvent),
    BlockBioFrontmerge(BlockBioFrontmergeFtraceEvent),
    BlockBioQueue(BlockBioQueueFtraceEvent),
    BlockBioRemap(BlockBioRemapFtraceEvent),
    BlockDirtyBuffer(BlockDirtyBufferFtraceEvent),
    BlockGetrq(BlockGetrqFtraceEvent),
    BlockPlug(BlockPlugFtraceEvent),
    BlockRqAbort(BlockRqAbortFtraceEvent),
    BlockRqComplete(BlockRqCompleteFtraceEvent),
    BlockRqInsert(BlockRqInsertFtraceEvent),
    BlockRqRemap(BlockRqRemapFtraceEvent),
    BlockRqRequeue(BlockRqRequeueFtraceEvent),
    BlockSleeprq(BlockSleeprqFtraceEvent),
    BlockSplit(BlockSplitFtraceEvent),
    BlockTouchBuffer(BlockTouchBufferFtraceEvent),
    BlockUnplug(BlockUnplugFtraceEvent),
    Ext4AllocDaBlocks(Ext4AllocDaBlocksFtraceEvent),
    Ext4AllocateBlocks(Ext4AllocateBlocksFtraceEvent),
    Ext4AllocateInode(Ext4AllocateInodeFtraceEvent),
    Ext4BeginOrderedTruncate(Ext4BeginOrderedTruncateFtraceEvent),
    Ext4CollapseRange(Ext4CollapseRangeFtraceEvent),
    Ext4DaReleaseSpace(Ext4DaReleaseSpaceFtraceEvent),
    Ext4DaReserveSpace(Ext4DaReserveSpaceFtraceEvent),
    Ext4DaUpdateReserveSpace(Ext4DaUpdateReserveSpaceFtraceEvent),
    Ext4DaWritePages(Ext4DaWritePagesFtraceEvent),
    Ext4DaWritePagesExtent(Ext4DaWritePagesExtentFtraceEvent),
    Ext4DirectIoEnter(Ext4DirectIoEnterFtraceEvent),
    Ext4DirectIoExit(Ext4DirectIoExitFtraceEvent),
    Ext4DiscardBlocks(Ext4DiscardBlocksFtraceEvent),
    Ext4DiscardPreallocations(Ext4DiscardPreallocationsFtraceEvent),
    Ext4DropInode(Ext4DropInodeFtraceEvent),
    Ext4EsCacheExtent(Ext4EsCacheExtentFtraceEvent),
    Ext4EsFindDelayedExtentRangeEnter(Ext4EsFindDelayedExtentRangeEnterFtraceEvent),
    Ext4EsFindDelayedExtentRangeExit(Ext4EsFindDelayedExtentRangeExitFtraceEvent),
    Ext4EsInsertExtent(Ext4EsInsertExtentFtraceEvent),
    Ext4EsLookupExtentEnter(Ext4EsLookupExtentEnterFtraceEvent),
    Ext4EsLookupExtentExit(Ext4EsLookupExtentExitFtraceEvent),
    Ext4EsRemoveExtent(Ext4EsRemoveExtentFtraceEvent),
    Ext4EsShrink(Ext4EsShrinkFtraceEvent),
    Ext4EsShrinkCount(Ext4EsShrinkCountFtraceEvent),
    Ext4EsShrinkScanEnter(Ext4EsShrinkScanEnterFtraceEvent),
    Ext4EsShrinkScanExit(Ext4EsShrinkScanExitFtraceEvent),
    Ext4EvictInode(Ext4EvictInodeFtraceEvent),
    Ext4ExtConvertToInitializedEnter(Ext4ExtConvertToInitializedEnterFtraceEvent),
    Ext4ExtConvertToInitializedFastpath(Ext4ExtConvertToInitializedFastpathFtraceEvent),
    Ext4ExtHandleUnwrittenExtents(Ext4ExtHandleUnwrittenExtentsFtraceEvent),
    Ext4ExtInCache(Ext4ExtInCacheFtraceEvent),
    Ext4ExtLoadExtent(Ext4ExtLoadExtentFtraceEvent),
    Ext4ExtMapBlocksEnter(Ext4ExtMapBlocksEnterFtraceEvent),
    Ext4ExtMapBlocksExit(Ext4ExtMapBlocksExitFtraceEvent),
    Ext4ExtPutInCache(Ext4ExtPutInCacheFtraceEvent),
    Ext4ExtRemoveSpace(Ext4ExtRemoveSpaceFtraceEvent),
    Ext4ExtRemoveSpaceDone(Ext4ExtRemoveSpaceDoneFtraceEvent),
    Ext4ExtRmIdx(Ext4ExtRmIdxFtraceEvent),
    Ext4ExtRmLeaf(Ext4ExtRmLeafFtraceEvent),
    Ext4ExtShowExtent(Ext4ExtShowExtentFtraceEvent),
    Ext4FallocateEnter(Ext4FallocateEnterFtraceEvent),
    Ext4FallocateExit(Ext4FallocateExitFtraceEvent),
    Ext4FindDelallocRange(Ext4FindDelallocRangeFtraceEvent),
    Ext4Forget(Ext4ForgetFtraceEvent),
    Ext4FreeBlocks(Ext4FreeBlocksFtraceEvent),
    Ext4FreeInode(Ext4FreeInodeFtraceEvent),
    Ext4GetImpliedClusterAllocExit(Ext4GetImpliedClusterAllocExitFtraceEvent),
    Ext4GetReservedClusterAlloc(Ext4GetReservedClusterAllocFtraceEvent),
    Ext4IndMapBlocksEnter(Ext4IndMapBlocksEnterFtraceEvent),
    Ext4IndMapBlocksExit(Ext4IndMapBlocksExitFtraceEvent),
    Ext4InsertRange(Ext4InsertRangeFtraceEvent),
    Ext4Invalidatepage(Ext4InvalidatepageFtraceEvent),
    Ext4JournalStart(Ext4JournalStartFtraceEvent),
    Ext4JournalStartReserved(Ext4JournalStartReservedFtraceEvent),
    Ext4JournalledInvalidatepage(Ext4JournalledInvalidatepageFtraceEvent),
    Ext4JournalledWriteEnd(Ext4JournalledWriteEndFtraceEvent),
    Ext4LoadInode(Ext4LoadInodeFtraceEvent),
    Ext4LoadInodeBitmap(Ext4LoadInodeBitmapFtraceEvent),
    Ext4MarkInodeDirty(Ext4MarkInodeDirtyFtraceEvent),
    Ext4MbBitmapLoad(Ext4MbBitmapLoadFtraceEvent),
    Ext4MbBuddyBitmapLoad(Ext4MbBuddyBitmapLoadFtraceEvent),
    Ext4MbDiscardPreallocations(Ext4MbDiscardPreallocationsFtraceEvent),
    Ext4MbNewGroupPa(Ext4MbNewGroupPaFtraceEvent),
    Ext4MbNewInodePa(Ext4MbNewInodePaFtraceEvent),
    Ext4MbReleaseGroupPa(Ext4MbReleaseGroupPaFtraceEvent),
    Ext4MbReleaseInodePa(Ext4MbReleaseInodePaFtraceEvent),
    Ext4MballocAlloc(Ext4MballocAllocFtraceEvent),
    Ext4MballocDiscard(Ext4MballocDiscardFtraceEvent),
    Ext4MballocFree(Ext4MballocFreeFtraceEvent),
    Ext4MballocPrealloc(Ext4MballocPreallocFtraceEvent),
    Ext4OtherInodeUpdateTime(Ext4OtherInodeUpdateTimeFtraceEvent),
    Ext4PunchHole(Ext4PunchHoleFtraceEvent),
    Ext4ReadBlockBitmapLoad(Ext4ReadBlockBitmapLoadFtraceEvent),
    Ext4Readpage(Ext4ReadpageFtraceEvent),
    Ext4Releasepage(Ext4ReleasepageFtraceEvent),
    Ext4RemoveBlocks(Ext4RemoveBlocksFtraceEvent),
    Ext4RequestBlocks(Ext4RequestBlocksFtraceEvent),
    Ext4RequestInode(Ext4RequestInodeFtraceEvent),
    Ext4SyncFs(Ext4SyncFsFtraceEvent),
    Ext4TrimAllFree(Ext4TrimAllFreeFtraceEvent),
    Ext4TrimExtent(Ext4TrimExtentFtraceEvent),
    Ext4TruncateEnter(Ext4TruncateEnterFtraceEvent),
    Ext4TruncateExit(Ext4TruncateExitFtraceEvent),
    Ext4UnlinkEnter(Ext4UnlinkEnterFtraceEvent),
    Ext4UnlinkExit(Ext4UnlinkExitFtraceEvent),
    Ext4WriteBegin(Ext4WriteBeginFtraceEvent),
    Ext4WriteEnd(Ext4WriteEndFtraceEvent),
    Ext4Writepage(Ext4WritepageFtraceEvent),
    Ext4Writepages(Ext4WritepagesFtraceEvent),
    Ext4WritepagesResult(Ext4WritepagesResultFtraceEvent),
    Ext4ZeroRange(Ext4ZeroRangeFtraceEvent),
    TaskNewtask(TaskNewtaskFtraceEvent),
    TaskRename(TaskRenameFtraceEvent),
    SchedProcessExec(SchedProcessExecFtraceEvent),
    SchedProcessExit(SchedProcessExitFtraceEvent),
    SchedProcessFork(SchedProcessForkFtraceEvent),
    SchedProcessFree(SchedProcessFreeFtraceEvent),
    SchedProcessHang(SchedProcessHangFtraceEvent),
    SchedProcessWait(SchedProcessWaitFtraceEvent),
    F2fsDoSubmitBio(F2fsDoSubmitBioFtraceEvent),
    F2fsEvictInode(F2fsEvictInodeFtraceEvent),
    F2fsFallocate(F2fsFallocateFtraceEvent),
    F2fsGetDataBlock(F2fsGetDataBlockFtraceEvent),
    F2fsGetVictim(F2fsGetVictimFtraceEvent),
    F2fsIget(F2fsIgetFtraceEvent),
    F2fsIgetExit(F2fsIgetExitFtraceEvent),
    F2fsNewInode(F2fsNewInodeFtraceEvent),
    F2fsReadpage(F2fsReadpageFtraceEvent),
    F2fsReserveNewBlock(F2fsReserveNewBlockFtraceEvent),
    F2fsSetPageDirty(F2fsSetPageDirtyFtraceEvent),
    F2fsSubmitWritePage(F2fsSubmitWritePageFtraceEvent),
    F2fsSyncFileEnter(F2fsSyncFileEnterFtraceEvent),
    F2fsSyncFileExit(F2fsSyncFileExitFtraceEvent),
    F2fsSyncFs(F2fsSyncFsFtraceEvent),
    F2fsTruncate(F2fsTruncateFtraceEvent),
    F2fsTruncateBlocksEnter(F2fsTruncateBlocksEnterFtraceEvent),
    F2fsTruncateBlocksExit(F2fsTruncateBlocksExitFtraceEvent),
    F2fsTruncateDataBlocksRange(F2fsTruncateDataBlocksRangeFtraceEvent),
    F2fsTruncateInodeBlocksEnter(F2fsTruncateInodeBlocksEnterFtraceEvent),
    F2fsTruncateInodeBlocksExit(F2fsTruncateInodeBlocksExitFtraceEvent),
    F2fsTruncateNode(F2fsTruncateNodeFtraceEvent),
    F2fsTruncateNodesEnter(F2fsTruncateNodesEnterFtraceEvent),
    F2fsTruncateNodesExit(F2fsTruncateNodesExitFtraceEvent),
    F2fsTruncatePartialNodes(F2fsTruncatePartialNodesFtraceEvent),
    F2fsUnlinkEnter(F2fsUnlinkEnterFtraceEvent),
    F2fsUnlinkExit(F2fsUnlinkExitFtraceEvent),
    F2fsVmPageMkwrite(F2fsVmPageMkwriteFtraceEvent),
    F2fsWriteBegin(F2fsWriteBeginFtraceEvent),
    F2fsWriteCheckpoint(F2fsWriteCheckpointFtraceEvent),
    F2fsWriteEnd(F2fsWriteEndFtraceEvent),
    AllocPagesIommuEnd(AllocPagesIommuEndFtraceEvent),
    AllocPagesIommuFail(AllocPagesIommuFailFtraceEvent),
    AllocPagesIommuStart(AllocPagesIommuStartFtraceEvent),
    AllocPagesSysEnd(AllocPagesSysEndFtraceEvent),
    AllocPagesSysFail(AllocPagesSysFailFtraceEvent),
    AllocPagesSysStart(AllocPagesSysStartFtraceEvent),
    DmaAllocContiguousRetry(DmaAllocContiguousRetryFtraceEvent),
    IommuMapRange(IommuMapRangeFtraceEvent),
    IommuSecPtblMapRangeEnd(IommuSecPtblMapRangeEndFtraceEvent),
    IommuSecPtblMapRangeStart(IommuSecPtblMapRangeStartFtraceEvent),
    IonAllocBufferEnd(IonAllocBufferEndFtraceEvent),
    IonAllocBufferFail(IonAllocBufferFailFtraceEvent),
    IonAllocBufferFallback(IonAllocBufferFallbackFtraceEvent),
    IonAllocBufferStart(IonAllocBufferStartFtraceEvent),
    IonCpAllocRetry(IonCpAllocRetryFtraceEvent),
    IonCpSecureBufferEnd(IonCpSecureBufferEndFtraceEvent),
    IonCpSecureBufferStart(IonCpSecureBufferStartFtraceEvent),
    IonPrefetching(IonPrefetchingFtraceEvent),
    IonSecureCmaAddToPoolEnd(IonSecureCmaAddToPoolEndFtraceEvent),
    IonSecureCmaAddToPoolStart(IonSecureCmaAddToPoolStartFtraceEvent),
    IonSecureCmaAllocateEnd(IonSecureCmaAllocateEndFtraceEvent),
    IonSecureCmaAllocateStart(IonSecureCmaAllocateStartFtraceEvent),
    IonSecureCmaShrinkPoolEnd(IonSecureCmaShrinkPoolEndFtraceEvent),
    IonSecureCmaShrinkPoolStart(IonSecureCmaShrinkPoolStartFtraceEvent),
    Kfree(KfreeFtraceEvent),
    Kmalloc(KmallocFtraceEvent),
    KmallocNode(KmallocNodeFtraceEvent),
    KmemCacheAlloc(KmemCacheAllocFtraceEvent),
    KmemCacheAllocNode(KmemCacheAllocNodeFtraceEvent),
    KmemCacheFree(KmemCacheFreeFtraceEvent),
    MigratePagesEnd(MigratePagesEndFtraceEvent),
    MigratePagesStart(MigratePagesStartFtraceEvent),
    MigrateRetry(MigrateRetryFtraceEvent),
    MmPageAlloc(MmPageAllocFtraceEvent),
    MmPageAllocExtfrag(MmPageAllocExtfragFtraceEvent),
    MmPageAllocZoneLocked(MmPageAllocZoneLockedFtraceEvent),
    MmPageFree(MmPageFreeFtraceEvent),
    MmPageFreeBatched(MmPageFreeBatchedFtraceEvent),
    MmPagePcpuDrain(MmPagePcpuDrainFtraceEvent),
    RssStat(RssStatFtraceEvent),
    IonHeapShrink(IonHeapShrinkFtraceEvent),
    IonHeapGrow(IonHeapGrowFtraceEvent),
    FenceInit(FenceInitFtraceEvent),
    FenceDestroy(FenceDestroyFtraceEvent),
    FenceEnableSignal(FenceEnableSignalFtraceEvent),
    FenceSignaled(FenceSignaledFtraceEvent),
    ClkEnable(ClkEnableFtraceEvent),
    ClkDisable(ClkDisableFtraceEvent),
    ClkSetRate(ClkSetRateFtraceEvent),
    BinderTransactionAllocBuf(BinderTransactionAllocBufFtraceEvent),
    SignalDeliver(SignalDeliverFtraceEvent),
    SignalGenerate(SignalGenerateFtraceEvent),
    OomScoreAdjUpdate(OomScoreAdjUpdateFtraceEvent),
    Generic(GenericFtraceEvent),
    MmEventRecord(MmEventRecordFtraceEvent),
    SysEnter(SysEnterFtraceEvent),
    SysExit(SysExitFtraceEvent),
    Zero(ZeroFtraceEvent),
    GpuFrequency(GpuFrequencyFtraceEvent),
    SdeTracingMarkWrite(SdeTracingMarkWriteFtraceEvent),
    MarkVictim(MarkVictimFtraceEvent),
    IonStat(IonStatFtraceEvent),
    IonBufferCreate(IonBufferCreateFtraceEvent),
    IonBufferDestroy(IonBufferDestroyFtraceEvent),
    ScmCallStart(ScmCallStartFtraceEvent),
    ScmCallEnd(ScmCallEndFtraceEvent),
    GpuMemTotal(GpuMemTotalFtraceEvent),
    ThermalTemperature(ThermalTemperatureFtraceEvent),
    CdevUpdate(CdevUpdateFtraceEvent),
    CpuhpExit(CpuhpExitFtraceEvent),
    CpuhpMultiEnter(CpuhpMultiEnterFtraceEvent),
    CpuhpEnter(CpuhpEnterFtraceEvent),
    CpuhpLatency(CpuhpLatencyFtraceEvent),
    FastrpcDmaStat(FastrpcDmaStatFtraceEvent),
    DpuTracingMarkWrite(DpuTracingMarkWriteFtraceEvent),
    G2dTracingMarkWrite(G2dTracingMarkWriteFtraceEvent),
    MaliTracingMarkWrite(MaliTracingMarkWriteFtraceEvent),
    DmaHeapStat(DmaHeapStatFtraceEvent),
    CpuhpPause(CpuhpPauseFtraceEvent),
    SchedPiSetprio(SchedPiSetprioFtraceEvent),
    SdeSdeEvtlog(SdeSdeEvtlogFtraceEvent),
    SdeSdePerfCalcCrtc(SdeSdePerfCalcCrtcFtraceEvent),
    SdeSdePerfCrtcUpdate(SdeSdePerfCrtcUpdateFtraceEvent),
    SdeSdePerfSetQosLuts(SdeSdePerfSetQosLutsFtraceEvent),
    SdeSdePerfUpdateBus(SdeSdePerfUpdateBusFtraceEvent),
    RssStatThrottled(RssStatThrottledFtraceEvent),
    NetifReceiveSkb(NetifReceiveSkbFtraceEvent),
    NetDevXmit(NetDevXmitFtraceEvent),
    InetSockSetState(InetSockSetStateFtraceEvent),
    TcpRetransmitSkb(TcpRetransmitSkbFtraceEvent),
    CrosEcSensorhubData(CrosEcSensorhubDataFtraceEvent),
    NapiGroReceiveEntry(NapiGroReceiveEntryFtraceEvent),
    NapiGroReceiveExit(NapiGroReceiveExitFtraceEvent),
    KfreeSkb(KfreeSkbFtraceEvent),
    KvmAccessFault(KvmAccessFaultFtraceEvent),
    KvmAckIrq(KvmAckIrqFtraceEvent),
    KvmAgeHva(KvmAgeHvaFtraceEvent),
    KvmAgePage(KvmAgePageFtraceEvent),
    KvmArmClearDebug(KvmArmClearDebugFtraceEvent),
    KvmArmSetDreg32(KvmArmSetDreg32FtraceEvent),
    KvmArmSetRegset(KvmArmSetRegsetFtraceEvent),
    KvmArmSetupDebug(KvmArmSetupDebugFtraceEvent),
    KvmEntry(KvmEntryFtraceEvent),
    KvmExit(KvmExitFtraceEvent),
    KvmFpu(KvmFpuFtraceEvent),
    KvmGetTimerMap(KvmGetTimerMapFtraceEvent),
    KvmGuestFault(KvmGuestFaultFtraceEvent),
    KvmHandleSysReg(KvmHandleSysRegFtraceEvent),
    KvmHvcArm64(KvmHvcArm64FtraceEvent),
    KvmIrqLine(KvmIrqLineFtraceEvent),
    KvmMmio(KvmMmioFtraceEvent),
    KvmMmioEmulate(KvmMmioEmulateFtraceEvent),
    KvmSetGuestDebug(KvmSetGuestDebugFtraceEvent),
    KvmSetIrq(KvmSetIrqFtraceEvent),
    KvmSetSpteHva(KvmSetSpteHvaFtraceEvent),
    KvmSetWayFlush(KvmSetWayFlushFtraceEvent),
    KvmSysAccess(KvmSysAccessFtraceEvent),
    KvmTestAgeHva(KvmTestAgeHvaFtraceEvent),
    KvmTimerEmulate(KvmTimerEmulateFtraceEvent),
    KvmTimerHrtimerExpire(KvmTimerHrtimerExpireFtraceEvent),
    KvmTimerRestoreState(KvmTimerRestoreStateFtraceEvent),
    KvmTimerSaveState(KvmTimerSaveStateFtraceEvent),
    KvmTimerUpdateIrq(KvmTimerUpdateIrqFtraceEvent),
    KvmToggleCache(KvmToggleCacheFtraceEvent),
    KvmUnmapHvaRange(KvmUnmapHvaRangeFtraceEvent),
    KvmUserspaceExit(KvmUserspaceExitFtraceEvent),
    KvmVcpuWakeup(KvmVcpuWakeupFtraceEvent),
    KvmWfxArm64(KvmWfxArm64FtraceEvent),
    TrapReg(TrapRegFtraceEvent),
    VgicUpdateIrqPending(VgicUpdateIrqPendingFtraceEvent),
    WakeupSourceActivate(WakeupSourceActivateFtraceEvent),
    WakeupSourceDeactivate(WakeupSourceDeactivateFtraceEvent),
    UfshcdCommand(UfshcdCommandFtraceEvent),
    UfshcdClkGating(UfshcdClkGatingFtraceEvent),
    Console(ConsoleFtraceEvent),
    DrmVblankEvent(DrmVblankEventFtraceEvent),
    DrmVblankEventDelivered(DrmVblankEventDeliveredFtraceEvent),
    DrmSchedJob(DrmSchedJobFtraceEvent),
    DrmRunJob(DrmRunJobFtraceEvent),
    DrmSchedProcessJob(DrmSchedProcessJobFtraceEvent),
    DmaFenceInit(DmaFenceInitFtraceEvent),
    DmaFenceEmit(DmaFenceEmitFtraceEvent),
    DmaFenceSignaled(DmaFenceSignaledFtraceEvent),
    DmaFenceWaitStart(DmaFenceWaitStartFtraceEvent),
    DmaFenceWaitEnd(DmaFenceWaitEndFtraceEvent),
    F2fsIostat(F2fsIostatFtraceEvent),
    F2fsIostatLatency(F2fsIostatLatencyFtraceEvent),
    SchedCpuUtilCfs(SchedCpuUtilCfsFtraceEvent),
    V4l2Qbuf(V4l2QbufFtraceEvent),
    V4l2Dqbuf(V4l2DqbufFtraceEvent),
    Vb2V4l2BufQueue(Vb2V4l2BufQueueFtraceEvent),
    Vb2V4l2BufDone(Vb2V4l2BufDoneFtraceEvent),
    Vb2V4l2Qbuf(Vb2V4l2QbufFtraceEvent),
    Vb2V4l2Dqbuf(Vb2V4l2DqbufFtraceEvent),
    DsiCmdFifoStatus(DsiCmdFifoStatusFtraceEvent),
    DsiRx(DsiRxFtraceEvent),
    DsiTx(DsiTxFtraceEvent),
    AndroidFsDatareadEnd(AndroidFsDatareadEndFtraceEvent),
    AndroidFsDatareadStart(AndroidFsDatareadStartFtraceEvent),
    AndroidFsDatawriteEnd(AndroidFsDatawriteEndFtraceEvent),
    AndroidFsDatawriteStart(AndroidFsDatawriteStartFtraceEvent),
    AndroidFsFsyncEnd(AndroidFsFsyncEndFtraceEvent),
    AndroidFsFsyncStart(AndroidFsFsyncStartFtraceEvent),
    FuncgraphEntry(FuncgraphEntryFtraceEvent),
    FuncgraphExit(FuncgraphExitFtraceEvent),
    VirtioVideoCmd(VirtioVideoCmdFtraceEvent),
    VirtioVideoCmdDone(VirtioVideoCmdDoneFtraceEvent),
    VirtioVideoResourceQueue(VirtioVideoResourceQueueFtraceEvent),
    VirtioVideoResourceQueueDone(VirtioVideoResourceQueueDoneFtraceEvent),
    MmShrinkSlabStart(MmShrinkSlabStartFtraceEvent),
    MmShrinkSlabEnd(MmShrinkSlabEndFtraceEvent),
    TrustySmc(TrustySmcFtraceEvent),
    TrustySmcDone(TrustySmcDoneFtraceEvent),
    TrustyStdCall32(TrustyStdCall32FtraceEvent),
    TrustyStdCall32Done(TrustyStdCall32DoneFtraceEvent),
    TrustyShareMemory(TrustyShareMemoryFtraceEvent),
    TrustyShareMemoryDone(TrustyShareMemoryDoneFtraceEvent),
    TrustyReclaimMemory(TrustyReclaimMemoryFtraceEvent),
    TrustyReclaimMemoryDone(TrustyReclaimMemoryDoneFtraceEvent),
    TrustyIrq(TrustyIrqFtraceEvent),
    TrustyIpcHandleEvent(TrustyIpcHandleEventFtraceEvent),
    TrustyIpcConnect(TrustyIpcConnectFtraceEvent),
    TrustyIpcConnectEnd(TrustyIpcConnectEndFtraceEvent),
    TrustyIpcWrite(TrustyIpcWriteFtraceEvent),
    TrustyIpcPoll(TrustyIpcPollFtraceEvent),
    TrustyIpcRead(TrustyIpcReadFtraceEvent),
    TrustyIpcReadEnd(TrustyIpcReadEndFtraceEvent),
    TrustyIpcRx(TrustyIpcRxFtraceEvent),
    TrustyEnqueueNop(TrustyEnqueueNopFtraceEvent),
    CmaAllocStart(CmaAllocStartFtraceEvent),
    CmaAllocInfo(CmaAllocInfoFtraceEvent),
    LwisTracingMarkWrite(LwisTracingMarkWriteFtraceEvent),
    VirtioGpuCmdQueue(VirtioGpuCmdQueueFtraceEvent),
    VirtioGpuCmdResponse(VirtioGpuCmdResponseFtraceEvent),
    MaliMaliKcpuCqsSet(MaliMaliKcpucqssetFtraceEvent),
    MaliMaliKcpuCqsWaitStart(MaliMaliKcpucqswaitstartFtraceEvent),
    MaliMaliKcpuCqsWaitEnd(MaliMaliKcpucqswaitendFtraceEvent),
    MaliMaliKcpuFenceSignal(MaliMaliKcpufencesignalFtraceEvent),
    MaliMaliKcpuFenceWaitStart(MaliMaliKcpufencewaitstartFtraceEvent),
    MaliMaliKcpuFenceWaitEnd(MaliMaliKcpufencewaitendFtraceEvent),
    HypEnter(HypEnterFtraceEvent),
    HypExit(HypExitFtraceEvent),
    HostHcall(HostHcallFtraceEvent),
    HostSmc(HostSmcFtraceEvent),
    HostMemAbort(HostMemAbortFtraceEvent),
    SuspendResumeMinimal(SuspendResumeMinimalFtraceEvent),
    MaliMaliCsfInterruptStart(MaliMaliCsfinterruptstartFtraceEvent),
    MaliMaliCsfInterruptEnd(MaliMaliCsfinterruptendFtraceEvent),
    SamsungTracingMarkWrite(SamsungTracingMarkWriteFtraceEvent),
    BinderCommand(BinderCommandFtraceEvent),
    BinderReturn(BinderReturnFtraceEvent),
    SchedSwitchWithCtrs(SchedSwitchWithCtrsFtraceEvent),
    GpuWorkPeriod(GpuWorkPeriodFtraceEvent),
    RpmStatus(RpmStatusFtraceEvent),
    PanelWriteGeneric(PanelWriteGenericFtraceEvent),
    SchedMigrateTask(SchedMigrateTaskFtraceEvent),
    DpuDsiCmdFifoStatus(DpuDsiCmdFifoStatusFtraceEvent),
    DpuDsiRx(DpuDsiRxFtraceEvent),
    DpuDsiTx(DpuDsiTxFtraceEvent),
    F2fsBackgroundGc(F2fsBackgroundGcFtraceEvent),
    F2fsGcBegin(F2fsGcBeginFtraceEvent),
    F2fsGcEnd(F2fsGcEndFtraceEvent),
    FastrpcDmaFree(FastrpcDmaFreeFtraceEvent),
    FastrpcDmaAlloc(FastrpcDmaAllocFtraceEvent),
    FastrpcDmaUnmap(FastrpcDmaUnmapFtraceEvent),
    FastrpcDmaMap(FastrpcDmaMapFtraceEvent),
    GoogleIccEvent(GoogleIccEventFtraceEvent),
    GoogleIrmEvent(GoogleIrmEventFtraceEvent),
    DevicePmCallbackStart(DevicePmCallbackStartFtraceEvent),
    DevicePmCallbackEnd(DevicePmCallbackEndFtraceEvent),
    ThermalExynosAcpmBulk(ThermalExynosAcpmBulkFtraceEvent),
    ThermalExynosAcpmHighOverhead(ThermalExynosAcpmHighOverheadFtraceEvent),
    DcvshFreq(DcvshFreqFtraceEvent),
    KgslGpuFrequency(KgslGpuFrequencyFtraceEvent),
    MaliMaliPmMcuHctlCoresDownScaleNotifyPend(MaliMaliPmmcuhctlcoresdownscalenotifypendFtraceEvent),
    MaliMaliPmMcuHctlCoresNotifyPend(MaliMaliPmmcuhctlcoresnotifypendFtraceEvent),
    MaliMaliPmMcuHctlCoreInactivePend(MaliMaliPmmcuhctlcoreinactivependFtraceEvent),
    MaliMaliPmMcuHctlMcuOnRecheck(MaliMaliPmmcuhctlmcuonrecheckFtraceEvent),
    MaliMaliPmMcuHctlShadersCoreOffPend(MaliMaliPmmcuhctlshaderscoreoffpendFtraceEvent),
    MaliMaliPmMcuHctlShadersPendOff(MaliMaliPmmcuhctlshaderspendoffFtraceEvent),
    MaliMaliPmMcuHctlShadersPendOn(MaliMaliPmmcuhctlshaderspendonFtraceEvent),
    MaliMaliPmMcuHctlShadersReadyOff(MaliMaliPmmcuhctlshadersreadyoffFtraceEvent),
    MaliMaliPmMcuInSleep(MaliMaliPmmcuinsleepFtraceEvent),
    MaliMaliPmMcuOff(MaliMaliPmmcuoffFtraceEvent),
    MaliMaliPmMcuOn(MaliMaliPmmcuonFtraceEvent),
    MaliMaliPmMcuOnCoreAttrUpdatePend(MaliMaliPmmcuoncoreattrupdatependFtraceEvent),
    MaliMaliPmMcuOnGlbReinitPend(MaliMaliPmmcuonglbreinitpendFtraceEvent),
    MaliMaliPmMcuOnHalt(MaliMaliPmmcuonhaltFtraceEvent),
    MaliMaliPmMcuOnHwcntDisable(MaliMaliPmmcuonhwcntdisableFtraceEvent),
    MaliMaliPmMcuOnHwcntEnable(MaliMaliPmmcuonhwcntenableFtraceEvent),
    MaliMaliPmMcuOnPendHalt(MaliMaliPmmcuonpendhaltFtraceEvent),
    MaliMaliPmMcuOnPendSleep(MaliMaliPmmcuonpendsleepFtraceEvent),
    MaliMaliPmMcuOnSleepInitiate(MaliMaliPmmcuonsleepinitiateFtraceEvent),
    MaliMaliPmMcuPendOff(MaliMaliPmmcupendoffFtraceEvent),
    MaliMaliPmMcuPendOnReload(MaliMaliPmmcupendonreloadFtraceEvent),
    MaliMaliPmMcuPowerDown(MaliMaliPmmcupowerdownFtraceEvent),
    MaliMaliPmMcuResetWait(MaliMaliPmmcuresetwaitFtraceEvent),

}

Variants

Print(PrintFtraceEvent)

SchedSwitch(SchedSwitchFtraceEvent)

CpuFrequency(CpuFrequencyFtraceEvent)

removed field with id 5; removed field with id 6; removed field with id 7; removed field with id 8; removed field with id 9; removed field with id 10;

CpuFrequencyLimits(CpuFrequencyLimitsFtraceEvent)

CpuIdle(CpuIdleFtraceEvent)

ClockEnable(ClockEnableFtraceEvent)

ClockDisable(ClockDisableFtraceEvent)

ClockSetRate(ClockSetRateFtraceEvent)

SchedWakeup(SchedWakeupFtraceEvent)

SchedBlockedReason(SchedBlockedReasonFtraceEvent)

SchedCpuHotplug(SchedCpuHotplugFtraceEvent)

SchedWaking(SchedWakingFtraceEvent)

IpiEntry(IpiEntryFtraceEvent)

IpiExit(IpiExitFtraceEvent)

IpiRaise(IpiRaiseFtraceEvent)

SoftirqEntry(SoftirqEntryFtraceEvent)

SoftirqExit(SoftirqExitFtraceEvent)

SoftirqRaise(SoftirqRaiseFtraceEvent)

I2cRead(I2cReadFtraceEvent)

I2cWrite(I2cWriteFtraceEvent)

I2cResult(I2cResultFtraceEvent)

I2cReply(I2cReplyFtraceEvent)

SmbusRead(SmbusReadFtraceEvent)

SmbusWrite(SmbusWriteFtraceEvent)

SmbusResult(SmbusResultFtraceEvent)

SmbusReply(SmbusReplyFtraceEvent)

LowmemoryKill(LowmemoryKillFtraceEvent)

IrqHandlerEntry(IrqHandlerEntryFtraceEvent)

IrqHandlerExit(IrqHandlerExitFtraceEvent)

SyncPt(SyncPtFtraceEvent)

SyncTimeline(SyncTimelineFtraceEvent)

SyncWait(SyncWaitFtraceEvent)

Ext4DaWriteBegin(Ext4DaWriteBeginFtraceEvent)

Ext4DaWriteEnd(Ext4DaWriteEndFtraceEvent)

Ext4SyncFileEnter(Ext4SyncFileEnterFtraceEvent)

Ext4SyncFileExit(Ext4SyncFileExitFtraceEvent)

BlockRqIssue(BlockRqIssueFtraceEvent)

MmVmscanDirectReclaimBegin(MmVmscanDirectReclaimBeginFtraceEvent)

MmVmscanDirectReclaimEnd(MmVmscanDirectReclaimEndFtraceEvent)

MmVmscanKswapdWake(MmVmscanKswapdWakeFtraceEvent)

MmVmscanKswapdSleep(MmVmscanKswapdSleepFtraceEvent)

BinderTransaction(BinderTransactionFtraceEvent)

BinderTransactionReceived(BinderTransactionReceivedFtraceEvent)

BinderSetPriority(BinderSetPriorityFtraceEvent)

BinderLock(BinderLockFtraceEvent)

BinderLocked(BinderLockedFtraceEvent)

BinderUnlock(BinderUnlockFtraceEvent)

WorkqueueActivateWork(WorkqueueActivateWorkFtraceEvent)

WorkqueueExecuteEnd(WorkqueueExecuteEndFtraceEvent)

WorkqueueExecuteStart(WorkqueueExecuteStartFtraceEvent)

WorkqueueQueueWork(WorkqueueQueueWorkFtraceEvent)

RegulatorDisable(RegulatorDisableFtraceEvent)

RegulatorDisableComplete(RegulatorDisableCompleteFtraceEvent)

RegulatorEnable(RegulatorEnableFtraceEvent)

RegulatorEnableComplete(RegulatorEnableCompleteFtraceEvent)

RegulatorEnableDelay(RegulatorEnableDelayFtraceEvent)

RegulatorSetVoltage(RegulatorSetVoltageFtraceEvent)

RegulatorSetVoltageComplete(RegulatorSetVoltageCompleteFtraceEvent)

CgroupAttachTask(CgroupAttachTaskFtraceEvent)

CgroupMkdir(CgroupMkdirFtraceEvent)

CgroupRemount(CgroupRemountFtraceEvent)

CgroupRmdir(CgroupRmdirFtraceEvent)

CgroupTransferTasks(CgroupTransferTasksFtraceEvent)

CgroupDestroyRoot(CgroupDestroyRootFtraceEvent)

CgroupRelease(CgroupReleaseFtraceEvent)

CgroupRename(CgroupRenameFtraceEvent)

CgroupSetupRoot(CgroupSetupRootFtraceEvent)

MdpCmdKickoff(MdpCmdKickoffFtraceEvent)

MdpCommit(MdpCommitFtraceEvent)

MdpPerfSetOt(MdpPerfSetOtFtraceEvent)

MdpSsppChange(MdpSsppChangeFtraceEvent)

TracingMarkWrite(TracingMarkWriteFtraceEvent)

MdpCmdPingpongDone(MdpCmdPingpongDoneFtraceEvent)

MdpCompareBw(MdpCompareBwFtraceEvent)

MdpPerfSetPanicLuts(MdpPerfSetPanicLutsFtraceEvent)

MdpSsppSet(MdpSsppSetFtraceEvent)

MdpCmdReadptrDone(MdpCmdReadptrDoneFtraceEvent)

MdpMisrCrc(MdpMisrCrcFtraceEvent)

MdpPerfSetQosLuts(MdpPerfSetQosLutsFtraceEvent)

MdpTraceCounter(MdpTraceCounterFtraceEvent)

MdpCmdReleaseBw(MdpCmdReleaseBwFtraceEvent)

MdpMixerUpdate(MdpMixerUpdateFtraceEvent)

MdpPerfSetWmLevels(MdpPerfSetWmLevelsFtraceEvent)

MdpVideoUnderrunDone(MdpVideoUnderrunDoneFtraceEvent)

MdpCmdWaitPingpong(MdpCmdWaitPingpongFtraceEvent)

MdpPerfPrefillCalc(MdpPerfPrefillCalcFtraceEvent)

MdpPerfUpdateBus(MdpPerfUpdateBusFtraceEvent)

RotatorBwAoAsContext(RotatorBwAoAsContextFtraceEvent)

MmFilemapAddToPageCache(MmFilemapAddToPageCacheFtraceEvent)

MmFilemapDeleteFromPageCache(MmFilemapDeleteFromPageCacheFtraceEvent)

MmCompactionBegin(MmCompactionBeginFtraceEvent)

MmCompactionDeferCompaction(MmCompactionDeferCompactionFtraceEvent)

MmCompactionDeferred(MmCompactionDeferredFtraceEvent)

MmCompactionDeferReset(MmCompactionDeferResetFtraceEvent)

MmCompactionEnd(MmCompactionEndFtraceEvent)

MmCompactionFinished(MmCompactionFinishedFtraceEvent)

MmCompactionIsolateFreepages(MmCompactionIsolateFreepagesFtraceEvent)

MmCompactionIsolateMigratepages(MmCompactionIsolateMigratepagesFtraceEvent)

MmCompactionKcompactdSleep(MmCompactionKcompactdSleepFtraceEvent)

MmCompactionKcompactdWake(MmCompactionKcompactdWakeFtraceEvent)

MmCompactionMigratepages(MmCompactionMigratepagesFtraceEvent)

MmCompactionSuitable(MmCompactionSuitableFtraceEvent)

MmCompactionTryToCompactPages(MmCompactionTryToCompactPagesFtraceEvent)

MmCompactionWakeupKcompactd(MmCompactionWakeupKcompactdFtraceEvent)

SuspendResume(SuspendResumeFtraceEvent)

SchedWakeupNew(SchedWakeupNewFtraceEvent)

BlockBioBackmerge(BlockBioBackmergeFtraceEvent)

BlockBioBounce(BlockBioBounceFtraceEvent)

BlockBioComplete(BlockBioCompleteFtraceEvent)

BlockBioFrontmerge(BlockBioFrontmergeFtraceEvent)

BlockBioQueue(BlockBioQueueFtraceEvent)

BlockBioRemap(BlockBioRemapFtraceEvent)

BlockDirtyBuffer(BlockDirtyBufferFtraceEvent)

BlockGetrq(BlockGetrqFtraceEvent)

BlockPlug(BlockPlugFtraceEvent)

BlockRqAbort(BlockRqAbortFtraceEvent)

BlockRqComplete(BlockRqCompleteFtraceEvent)

BlockRqInsert(BlockRqInsertFtraceEvent)

BlockRqRemap(BlockRqRemapFtraceEvent)

removed field with id 127;

BlockRqRequeue(BlockRqRequeueFtraceEvent)

BlockSleeprq(BlockSleeprqFtraceEvent)

BlockSplit(BlockSplitFtraceEvent)

BlockTouchBuffer(BlockTouchBufferFtraceEvent)

BlockUnplug(BlockUnplugFtraceEvent)

Ext4AllocDaBlocks(Ext4AllocDaBlocksFtraceEvent)

Ext4AllocateBlocks(Ext4AllocateBlocksFtraceEvent)

Ext4AllocateInode(Ext4AllocateInodeFtraceEvent)

Ext4BeginOrderedTruncate(Ext4BeginOrderedTruncateFtraceEvent)

Ext4CollapseRange(Ext4CollapseRangeFtraceEvent)

Ext4DaReleaseSpace(Ext4DaReleaseSpaceFtraceEvent)

Ext4DaReserveSpace(Ext4DaReserveSpaceFtraceEvent)

Ext4DaUpdateReserveSpace(Ext4DaUpdateReserveSpaceFtraceEvent)

Ext4DaWritePages(Ext4DaWritePagesFtraceEvent)

Ext4DaWritePagesExtent(Ext4DaWritePagesExtentFtraceEvent)

Ext4DirectIoEnter(Ext4DirectIoEnterFtraceEvent)

Ext4DirectIoExit(Ext4DirectIoExitFtraceEvent)

Ext4DiscardBlocks(Ext4DiscardBlocksFtraceEvent)

Ext4DiscardPreallocations(Ext4DiscardPreallocationsFtraceEvent)

Ext4DropInode(Ext4DropInodeFtraceEvent)

Ext4EsCacheExtent(Ext4EsCacheExtentFtraceEvent)

Ext4EsFindDelayedExtentRangeEnter(Ext4EsFindDelayedExtentRangeEnterFtraceEvent)

Ext4EsFindDelayedExtentRangeExit(Ext4EsFindDelayedExtentRangeExitFtraceEvent)

Ext4EsInsertExtent(Ext4EsInsertExtentFtraceEvent)

Ext4EsLookupExtentEnter(Ext4EsLookupExtentEnterFtraceEvent)

Ext4EsLookupExtentExit(Ext4EsLookupExtentExitFtraceEvent)

Ext4EsRemoveExtent(Ext4EsRemoveExtentFtraceEvent)

Ext4EsShrink(Ext4EsShrinkFtraceEvent)

Ext4EsShrinkCount(Ext4EsShrinkCountFtraceEvent)

Ext4EsShrinkScanEnter(Ext4EsShrinkScanEnterFtraceEvent)

Ext4EsShrinkScanExit(Ext4EsShrinkScanExitFtraceEvent)

Ext4EvictInode(Ext4EvictInodeFtraceEvent)

Ext4ExtConvertToInitializedEnter(Ext4ExtConvertToInitializedEnterFtraceEvent)

Ext4ExtConvertToInitializedFastpath(Ext4ExtConvertToInitializedFastpathFtraceEvent)

Ext4ExtHandleUnwrittenExtents(Ext4ExtHandleUnwrittenExtentsFtraceEvent)

Ext4ExtInCache(Ext4ExtInCacheFtraceEvent)

Ext4ExtLoadExtent(Ext4ExtLoadExtentFtraceEvent)

Ext4ExtMapBlocksEnter(Ext4ExtMapBlocksEnterFtraceEvent)

Ext4ExtMapBlocksExit(Ext4ExtMapBlocksExitFtraceEvent)

Ext4ExtPutInCache(Ext4ExtPutInCacheFtraceEvent)

Ext4ExtRemoveSpace(Ext4ExtRemoveSpaceFtraceEvent)

Ext4ExtRemoveSpaceDone(Ext4ExtRemoveSpaceDoneFtraceEvent)

Ext4ExtRmIdx(Ext4ExtRmIdxFtraceEvent)

Ext4ExtRmLeaf(Ext4ExtRmLeafFtraceEvent)

Ext4ExtShowExtent(Ext4ExtShowExtentFtraceEvent)

Ext4FallocateEnter(Ext4FallocateEnterFtraceEvent)

Ext4FallocateExit(Ext4FallocateExitFtraceEvent)

Ext4FindDelallocRange(Ext4FindDelallocRangeFtraceEvent)

Ext4Forget(Ext4ForgetFtraceEvent)

Ext4FreeBlocks(Ext4FreeBlocksFtraceEvent)

Ext4FreeInode(Ext4FreeInodeFtraceEvent)

Ext4GetImpliedClusterAllocExit(Ext4GetImpliedClusterAllocExitFtraceEvent)

Ext4GetReservedClusterAlloc(Ext4GetReservedClusterAllocFtraceEvent)

Ext4IndMapBlocksEnter(Ext4IndMapBlocksEnterFtraceEvent)

Ext4IndMapBlocksExit(Ext4IndMapBlocksExitFtraceEvent)

Ext4InsertRange(Ext4InsertRangeFtraceEvent)

Ext4Invalidatepage(Ext4InvalidatepageFtraceEvent)

Ext4JournalStart(Ext4JournalStartFtraceEvent)

Ext4JournalStartReserved(Ext4JournalStartReservedFtraceEvent)

Ext4JournalledInvalidatepage(Ext4JournalledInvalidatepageFtraceEvent)

Ext4JournalledWriteEnd(Ext4JournalledWriteEndFtraceEvent)

Ext4LoadInode(Ext4LoadInodeFtraceEvent)

Ext4LoadInodeBitmap(Ext4LoadInodeBitmapFtraceEvent)

Ext4MarkInodeDirty(Ext4MarkInodeDirtyFtraceEvent)

Ext4MbBitmapLoad(Ext4MbBitmapLoadFtraceEvent)

Ext4MbBuddyBitmapLoad(Ext4MbBuddyBitmapLoadFtraceEvent)

Ext4MbDiscardPreallocations(Ext4MbDiscardPreallocationsFtraceEvent)

Ext4MbNewGroupPa(Ext4MbNewGroupPaFtraceEvent)

Ext4MbNewInodePa(Ext4MbNewInodePaFtraceEvent)

Ext4MbReleaseGroupPa(Ext4MbReleaseGroupPaFtraceEvent)

Ext4MbReleaseInodePa(Ext4MbReleaseInodePaFtraceEvent)

Ext4MballocAlloc(Ext4MballocAllocFtraceEvent)

Ext4MballocDiscard(Ext4MballocDiscardFtraceEvent)

Ext4MballocFree(Ext4MballocFreeFtraceEvent)

Ext4MballocPrealloc(Ext4MballocPreallocFtraceEvent)

Ext4OtherInodeUpdateTime(Ext4OtherInodeUpdateTimeFtraceEvent)

Ext4PunchHole(Ext4PunchHoleFtraceEvent)

Ext4ReadBlockBitmapLoad(Ext4ReadBlockBitmapLoadFtraceEvent)

Ext4Readpage(Ext4ReadpageFtraceEvent)

Ext4Releasepage(Ext4ReleasepageFtraceEvent)

Ext4RemoveBlocks(Ext4RemoveBlocksFtraceEvent)

Ext4RequestBlocks(Ext4RequestBlocksFtraceEvent)

Ext4RequestInode(Ext4RequestInodeFtraceEvent)

Ext4SyncFs(Ext4SyncFsFtraceEvent)

Ext4TrimAllFree(Ext4TrimAllFreeFtraceEvent)

Ext4TrimExtent(Ext4TrimExtentFtraceEvent)

Ext4TruncateEnter(Ext4TruncateEnterFtraceEvent)

Ext4TruncateExit(Ext4TruncateExitFtraceEvent)

Ext4UnlinkEnter(Ext4UnlinkEnterFtraceEvent)

Ext4UnlinkExit(Ext4UnlinkExitFtraceEvent)

Ext4WriteBegin(Ext4WriteBeginFtraceEvent)

Ext4WriteEnd(Ext4WriteEndFtraceEvent)

removed field with id 220; removed field with id 221; removed field with id 222; removed field with id 223; removed field with id 224; removed field with id 225; removed field with id 226; removed field with id 227; removed field with id 228; removed field with id 229;

Ext4Writepage(Ext4WritepageFtraceEvent)

Ext4Writepages(Ext4WritepagesFtraceEvent)

Ext4WritepagesResult(Ext4WritepagesResultFtraceEvent)

Ext4ZeroRange(Ext4ZeroRangeFtraceEvent)

TaskNewtask(TaskNewtaskFtraceEvent)

TaskRename(TaskRenameFtraceEvent)

SchedProcessExec(SchedProcessExecFtraceEvent)

SchedProcessExit(SchedProcessExitFtraceEvent)

SchedProcessFork(SchedProcessForkFtraceEvent)

SchedProcessFree(SchedProcessFreeFtraceEvent)

SchedProcessHang(SchedProcessHangFtraceEvent)

SchedProcessWait(SchedProcessWaitFtraceEvent)

F2fsDoSubmitBio(F2fsDoSubmitBioFtraceEvent)

F2fsEvictInode(F2fsEvictInodeFtraceEvent)

F2fsFallocate(F2fsFallocateFtraceEvent)

F2fsGetDataBlock(F2fsGetDataBlockFtraceEvent)

F2fsGetVictim(F2fsGetVictimFtraceEvent)

F2fsIget(F2fsIgetFtraceEvent)

F2fsIgetExit(F2fsIgetExitFtraceEvent)

F2fsNewInode(F2fsNewInodeFtraceEvent)

F2fsReadpage(F2fsReadpageFtraceEvent)

F2fsReserveNewBlock(F2fsReserveNewBlockFtraceEvent)

F2fsSetPageDirty(F2fsSetPageDirtyFtraceEvent)

F2fsSubmitWritePage(F2fsSubmitWritePageFtraceEvent)

F2fsSyncFileEnter(F2fsSyncFileEnterFtraceEvent)

F2fsSyncFileExit(F2fsSyncFileExitFtraceEvent)

F2fsSyncFs(F2fsSyncFsFtraceEvent)

F2fsTruncate(F2fsTruncateFtraceEvent)

F2fsTruncateBlocksEnter(F2fsTruncateBlocksEnterFtraceEvent)

F2fsTruncateBlocksExit(F2fsTruncateBlocksExitFtraceEvent)

F2fsTruncateDataBlocksRange(F2fsTruncateDataBlocksRangeFtraceEvent)

F2fsTruncateInodeBlocksEnter(F2fsTruncateInodeBlocksEnterFtraceEvent)

F2fsTruncateInodeBlocksExit(F2fsTruncateInodeBlocksExitFtraceEvent)

F2fsTruncateNode(F2fsTruncateNodeFtraceEvent)

F2fsTruncateNodesEnter(F2fsTruncateNodesEnterFtraceEvent)

F2fsTruncateNodesExit(F2fsTruncateNodesExitFtraceEvent)

F2fsTruncatePartialNodes(F2fsTruncatePartialNodesFtraceEvent)

F2fsUnlinkEnter(F2fsUnlinkEnterFtraceEvent)

F2fsUnlinkExit(F2fsUnlinkExitFtraceEvent)

F2fsVmPageMkwrite(F2fsVmPageMkwriteFtraceEvent)

F2fsWriteBegin(F2fsWriteBeginFtraceEvent)

F2fsWriteCheckpoint(F2fsWriteCheckpointFtraceEvent)

F2fsWriteEnd(F2fsWriteEndFtraceEvent)

AllocPagesIommuEnd(AllocPagesIommuEndFtraceEvent)

AllocPagesIommuFail(AllocPagesIommuFailFtraceEvent)

AllocPagesIommuStart(AllocPagesIommuStartFtraceEvent)

AllocPagesSysEnd(AllocPagesSysEndFtraceEvent)

AllocPagesSysFail(AllocPagesSysFailFtraceEvent)

AllocPagesSysStart(AllocPagesSysStartFtraceEvent)

DmaAllocContiguousRetry(DmaAllocContiguousRetryFtraceEvent)

IommuMapRange(IommuMapRangeFtraceEvent)

IommuSecPtblMapRangeEnd(IommuSecPtblMapRangeEndFtraceEvent)

IommuSecPtblMapRangeStart(IommuSecPtblMapRangeStartFtraceEvent)

IonAllocBufferEnd(IonAllocBufferEndFtraceEvent)

IonAllocBufferFail(IonAllocBufferFailFtraceEvent)

IonAllocBufferFallback(IonAllocBufferFallbackFtraceEvent)

IonAllocBufferStart(IonAllocBufferStartFtraceEvent)

IonCpAllocRetry(IonCpAllocRetryFtraceEvent)

IonCpSecureBufferEnd(IonCpSecureBufferEndFtraceEvent)

IonCpSecureBufferStart(IonCpSecureBufferStartFtraceEvent)

IonPrefetching(IonPrefetchingFtraceEvent)

IonSecureCmaAddToPoolEnd(IonSecureCmaAddToPoolEndFtraceEvent)

IonSecureCmaAddToPoolStart(IonSecureCmaAddToPoolStartFtraceEvent)

IonSecureCmaAllocateEnd(IonSecureCmaAllocateEndFtraceEvent)

IonSecureCmaAllocateStart(IonSecureCmaAllocateStartFtraceEvent)

IonSecureCmaShrinkPoolEnd(IonSecureCmaShrinkPoolEndFtraceEvent)

IonSecureCmaShrinkPoolStart(IonSecureCmaShrinkPoolStartFtraceEvent)

Kfree(KfreeFtraceEvent)

Kmalloc(KmallocFtraceEvent)

KmallocNode(KmallocNodeFtraceEvent)

KmemCacheAlloc(KmemCacheAllocFtraceEvent)

KmemCacheAllocNode(KmemCacheAllocNodeFtraceEvent)

KmemCacheFree(KmemCacheFreeFtraceEvent)

MigratePagesEnd(MigratePagesEndFtraceEvent)

MigratePagesStart(MigratePagesStartFtraceEvent)

MigrateRetry(MigrateRetryFtraceEvent)

MmPageAlloc(MmPageAllocFtraceEvent)

MmPageAllocExtfrag(MmPageAllocExtfragFtraceEvent)

MmPageAllocZoneLocked(MmPageAllocZoneLockedFtraceEvent)

MmPageFree(MmPageFreeFtraceEvent)

MmPageFreeBatched(MmPageFreeBatchedFtraceEvent)

MmPagePcpuDrain(MmPagePcpuDrainFtraceEvent)

RssStat(RssStatFtraceEvent)

IonHeapShrink(IonHeapShrinkFtraceEvent)

IonHeapGrow(IonHeapGrowFtraceEvent)

FenceInit(FenceInitFtraceEvent)

FenceDestroy(FenceDestroyFtraceEvent)

FenceEnableSignal(FenceEnableSignalFtraceEvent)

FenceSignaled(FenceSignaledFtraceEvent)

ClkEnable(ClkEnableFtraceEvent)

ClkDisable(ClkDisableFtraceEvent)

ClkSetRate(ClkSetRateFtraceEvent)

BinderTransactionAllocBuf(BinderTransactionAllocBufFtraceEvent)

SignalDeliver(SignalDeliverFtraceEvent)

SignalGenerate(SignalGenerateFtraceEvent)

OomScoreAdjUpdate(OomScoreAdjUpdateFtraceEvent)

Generic(GenericFtraceEvent)

MmEventRecord(MmEventRecordFtraceEvent)

SysEnter(SysEnterFtraceEvent)

SysExit(SysExitFtraceEvent)

Zero(ZeroFtraceEvent)

GpuFrequency(GpuFrequencyFtraceEvent)

SdeTracingMarkWrite(SdeTracingMarkWriteFtraceEvent)

MarkVictim(MarkVictimFtraceEvent)

IonStat(IonStatFtraceEvent)

IonBufferCreate(IonBufferCreateFtraceEvent)

IonBufferDestroy(IonBufferDestroyFtraceEvent)

ScmCallStart(ScmCallStartFtraceEvent)

ScmCallEnd(ScmCallEndFtraceEvent)

GpuMemTotal(GpuMemTotalFtraceEvent)

ThermalTemperature(ThermalTemperatureFtraceEvent)

CdevUpdate(CdevUpdateFtraceEvent)

CpuhpExit(CpuhpExitFtraceEvent)

CpuhpMultiEnter(CpuhpMultiEnterFtraceEvent)

CpuhpEnter(CpuhpEnterFtraceEvent)

CpuhpLatency(CpuhpLatencyFtraceEvent)

FastrpcDmaStat(FastrpcDmaStatFtraceEvent)

DpuTracingMarkWrite(DpuTracingMarkWriteFtraceEvent)

G2dTracingMarkWrite(G2dTracingMarkWriteFtraceEvent)

MaliTracingMarkWrite(MaliTracingMarkWriteFtraceEvent)

DmaHeapStat(DmaHeapStatFtraceEvent)

CpuhpPause(CpuhpPauseFtraceEvent)

SchedPiSetprio(SchedPiSetprioFtraceEvent)

SdeSdeEvtlog(SdeSdeEvtlogFtraceEvent)

SdeSdePerfCalcCrtc(SdeSdePerfCalcCrtcFtraceEvent)

SdeSdePerfCrtcUpdate(SdeSdePerfCrtcUpdateFtraceEvent)

SdeSdePerfSetQosLuts(SdeSdePerfSetQosLutsFtraceEvent)

SdeSdePerfUpdateBus(SdeSdePerfUpdateBusFtraceEvent)

RssStatThrottled(RssStatThrottledFtraceEvent)

NetifReceiveSkb(NetifReceiveSkbFtraceEvent)

NetDevXmit(NetDevXmitFtraceEvent)

InetSockSetState(InetSockSetStateFtraceEvent)

TcpRetransmitSkb(TcpRetransmitSkbFtraceEvent)

CrosEcSensorhubData(CrosEcSensorhubDataFtraceEvent)

NapiGroReceiveEntry(NapiGroReceiveEntryFtraceEvent)

NapiGroReceiveExit(NapiGroReceiveExitFtraceEvent)

KfreeSkb(KfreeSkbFtraceEvent)

KvmAccessFault(KvmAccessFaultFtraceEvent)

KvmAckIrq(KvmAckIrqFtraceEvent)

KvmAgeHva(KvmAgeHvaFtraceEvent)

KvmAgePage(KvmAgePageFtraceEvent)

KvmArmClearDebug(KvmArmClearDebugFtraceEvent)

KvmArmSetDreg32(KvmArmSetDreg32FtraceEvent)

KvmArmSetRegset(KvmArmSetRegsetFtraceEvent)

KvmArmSetupDebug(KvmArmSetupDebugFtraceEvent)

KvmEntry(KvmEntryFtraceEvent)

KvmExit(KvmExitFtraceEvent)

KvmFpu(KvmFpuFtraceEvent)

KvmGetTimerMap(KvmGetTimerMapFtraceEvent)

KvmGuestFault(KvmGuestFaultFtraceEvent)

KvmHandleSysReg(KvmHandleSysRegFtraceEvent)

KvmHvcArm64(KvmHvcArm64FtraceEvent)

KvmIrqLine(KvmIrqLineFtraceEvent)

KvmMmio(KvmMmioFtraceEvent)

KvmMmioEmulate(KvmMmioEmulateFtraceEvent)

KvmSetGuestDebug(KvmSetGuestDebugFtraceEvent)

KvmSetIrq(KvmSetIrqFtraceEvent)

KvmSetSpteHva(KvmSetSpteHvaFtraceEvent)

KvmSetWayFlush(KvmSetWayFlushFtraceEvent)

KvmSysAccess(KvmSysAccessFtraceEvent)

KvmTestAgeHva(KvmTestAgeHvaFtraceEvent)

KvmTimerEmulate(KvmTimerEmulateFtraceEvent)

KvmTimerHrtimerExpire(KvmTimerHrtimerExpireFtraceEvent)

KvmTimerRestoreState(KvmTimerRestoreStateFtraceEvent)

KvmTimerSaveState(KvmTimerSaveStateFtraceEvent)

KvmTimerUpdateIrq(KvmTimerUpdateIrqFtraceEvent)

KvmToggleCache(KvmToggleCacheFtraceEvent)

KvmUnmapHvaRange(KvmUnmapHvaRangeFtraceEvent)

KvmUserspaceExit(KvmUserspaceExitFtraceEvent)

KvmVcpuWakeup(KvmVcpuWakeupFtraceEvent)

KvmWfxArm64(KvmWfxArm64FtraceEvent)

TrapReg(TrapRegFtraceEvent)

VgicUpdateIrqPending(VgicUpdateIrqPendingFtraceEvent)

WakeupSourceActivate(WakeupSourceActivateFtraceEvent)

WakeupSourceDeactivate(WakeupSourceDeactivateFtraceEvent)

UfshcdCommand(UfshcdCommandFtraceEvent)

UfshcdClkGating(UfshcdClkGatingFtraceEvent)

Console(ConsoleFtraceEvent)

DrmVblankEvent(DrmVblankEventFtraceEvent)

DrmVblankEventDelivered(DrmVblankEventDeliveredFtraceEvent)

DrmSchedJob(DrmSchedJobFtraceEvent)

DrmRunJob(DrmRunJobFtraceEvent)

DrmSchedProcessJob(DrmSchedProcessJobFtraceEvent)

DmaFenceInit(DmaFenceInitFtraceEvent)

DmaFenceEmit(DmaFenceEmitFtraceEvent)

DmaFenceSignaled(DmaFenceSignaledFtraceEvent)

DmaFenceWaitStart(DmaFenceWaitStartFtraceEvent)

DmaFenceWaitEnd(DmaFenceWaitEndFtraceEvent)

F2fsIostat(F2fsIostatFtraceEvent)

F2fsIostatLatency(F2fsIostatLatencyFtraceEvent)

SchedCpuUtilCfs(SchedCpuUtilCfsFtraceEvent)

V4l2Qbuf(V4l2QbufFtraceEvent)

V4l2Dqbuf(V4l2DqbufFtraceEvent)

Vb2V4l2BufQueue(Vb2V4l2BufQueueFtraceEvent)

Vb2V4l2BufDone(Vb2V4l2BufDoneFtraceEvent)

Vb2V4l2Qbuf(Vb2V4l2QbufFtraceEvent)

Vb2V4l2Dqbuf(Vb2V4l2DqbufFtraceEvent)

DsiCmdFifoStatus(DsiCmdFifoStatusFtraceEvent)

DsiRx(DsiRxFtraceEvent)

DsiTx(DsiTxFtraceEvent)

AndroidFsDatareadEnd(AndroidFsDatareadEndFtraceEvent)

AndroidFsDatareadStart(AndroidFsDatareadStartFtraceEvent)

AndroidFsDatawriteEnd(AndroidFsDatawriteEndFtraceEvent)

AndroidFsDatawriteStart(AndroidFsDatawriteStartFtraceEvent)

AndroidFsFsyncEnd(AndroidFsFsyncEndFtraceEvent)

AndroidFsFsyncStart(AndroidFsFsyncStartFtraceEvent)

FuncgraphEntry(FuncgraphEntryFtraceEvent)

FuncgraphExit(FuncgraphExitFtraceEvent)

VirtioVideoCmd(VirtioVideoCmdFtraceEvent)

VirtioVideoCmdDone(VirtioVideoCmdDoneFtraceEvent)

VirtioVideoResourceQueue(VirtioVideoResourceQueueFtraceEvent)

VirtioVideoResourceQueueDone(VirtioVideoResourceQueueDoneFtraceEvent)

MmShrinkSlabStart(MmShrinkSlabStartFtraceEvent)

MmShrinkSlabEnd(MmShrinkSlabEndFtraceEvent)

TrustySmc(TrustySmcFtraceEvent)

TrustySmcDone(TrustySmcDoneFtraceEvent)

TrustyStdCall32(TrustyStdCall32FtraceEvent)

TrustyStdCall32Done(TrustyStdCall32DoneFtraceEvent)

TrustyShareMemory(TrustyShareMemoryFtraceEvent)

TrustyShareMemoryDone(TrustyShareMemoryDoneFtraceEvent)

TrustyReclaimMemory(TrustyReclaimMemoryFtraceEvent)

TrustyReclaimMemoryDone(TrustyReclaimMemoryDoneFtraceEvent)

TrustyIrq(TrustyIrqFtraceEvent)

TrustyIpcHandleEvent(TrustyIpcHandleEventFtraceEvent)

TrustyIpcConnect(TrustyIpcConnectFtraceEvent)

TrustyIpcConnectEnd(TrustyIpcConnectEndFtraceEvent)

TrustyIpcWrite(TrustyIpcWriteFtraceEvent)

TrustyIpcPoll(TrustyIpcPollFtraceEvent)

TrustyIpcRead(TrustyIpcReadFtraceEvent)

removed field with id 459;

TrustyIpcReadEnd(TrustyIpcReadEndFtraceEvent)

TrustyIpcRx(TrustyIpcRxFtraceEvent)

TrustyEnqueueNop(TrustyEnqueueNopFtraceEvent)

removed field with id 463;

CmaAllocStart(CmaAllocStartFtraceEvent)

CmaAllocInfo(CmaAllocInfoFtraceEvent)

LwisTracingMarkWrite(LwisTracingMarkWriteFtraceEvent)

VirtioGpuCmdQueue(VirtioGpuCmdQueueFtraceEvent)

VirtioGpuCmdResponse(VirtioGpuCmdResponseFtraceEvent)

MaliMaliKcpuCqsSet(MaliMaliKcpucqssetFtraceEvent)

MaliMaliKcpuCqsWaitStart(MaliMaliKcpucqswaitstartFtraceEvent)

MaliMaliKcpuCqsWaitEnd(MaliMaliKcpucqswaitendFtraceEvent)

MaliMaliKcpuFenceSignal(MaliMaliKcpufencesignalFtraceEvent)

MaliMaliKcpuFenceWaitStart(MaliMaliKcpufencewaitstartFtraceEvent)

MaliMaliKcpuFenceWaitEnd(MaliMaliKcpufencewaitendFtraceEvent)

HypEnter(HypEnterFtraceEvent)

HypExit(HypExitFtraceEvent)

HostHcall(HostHcallFtraceEvent)

HostSmc(HostSmcFtraceEvent)

HostMemAbort(HostMemAbortFtraceEvent)

SuspendResumeMinimal(SuspendResumeMinimalFtraceEvent)

MaliMaliCsfInterruptStart(MaliMaliCsfinterruptstartFtraceEvent)

MaliMaliCsfInterruptEnd(MaliMaliCsfinterruptendFtraceEvent)

SamsungTracingMarkWrite(SamsungTracingMarkWriteFtraceEvent)

BinderCommand(BinderCommandFtraceEvent)

BinderReturn(BinderReturnFtraceEvent)

SchedSwitchWithCtrs(SchedSwitchWithCtrsFtraceEvent)

GpuWorkPeriod(GpuWorkPeriodFtraceEvent)

RpmStatus(RpmStatusFtraceEvent)

PanelWriteGeneric(PanelWriteGenericFtraceEvent)

SchedMigrateTask(SchedMigrateTaskFtraceEvent)

DpuDsiCmdFifoStatus(DpuDsiCmdFifoStatusFtraceEvent)

DpuDsiRx(DpuDsiRxFtraceEvent)

DpuDsiTx(DpuDsiTxFtraceEvent)

F2fsBackgroundGc(F2fsBackgroundGcFtraceEvent)

F2fsGcBegin(F2fsGcBeginFtraceEvent)

F2fsGcEnd(F2fsGcEndFtraceEvent)

FastrpcDmaFree(FastrpcDmaFreeFtraceEvent)

FastrpcDmaAlloc(FastrpcDmaAllocFtraceEvent)

FastrpcDmaUnmap(FastrpcDmaUnmapFtraceEvent)

FastrpcDmaMap(FastrpcDmaMapFtraceEvent)

GoogleIccEvent(GoogleIccEventFtraceEvent)

GoogleIrmEvent(GoogleIrmEventFtraceEvent)

DevicePmCallbackStart(DevicePmCallbackStartFtraceEvent)

DevicePmCallbackEnd(DevicePmCallbackEndFtraceEvent)

ThermalExynosAcpmBulk(ThermalExynosAcpmBulkFtraceEvent)

ThermalExynosAcpmHighOverhead(ThermalExynosAcpmHighOverheadFtraceEvent)

DcvshFreq(DcvshFreqFtraceEvent)

KgslGpuFrequency(KgslGpuFrequencyFtraceEvent)

MaliMaliPmMcuHctlCoresDownScaleNotifyPend(MaliMaliPmmcuhctlcoresdownscalenotifypendFtraceEvent)

MaliMaliPmMcuHctlCoresNotifyPend(MaliMaliPmmcuhctlcoresnotifypendFtraceEvent)

MaliMaliPmMcuHctlCoreInactivePend(MaliMaliPmmcuhctlcoreinactivependFtraceEvent)

MaliMaliPmMcuHctlMcuOnRecheck(MaliMaliPmmcuhctlmcuonrecheckFtraceEvent)

MaliMaliPmMcuHctlShadersCoreOffPend(MaliMaliPmmcuhctlshaderscoreoffpendFtraceEvent)

MaliMaliPmMcuHctlShadersPendOff(MaliMaliPmmcuhctlshaderspendoffFtraceEvent)

MaliMaliPmMcuHctlShadersPendOn(MaliMaliPmmcuhctlshaderspendonFtraceEvent)

MaliMaliPmMcuHctlShadersReadyOff(MaliMaliPmmcuhctlshadersreadyoffFtraceEvent)

MaliMaliPmMcuInSleep(MaliMaliPmmcuinsleepFtraceEvent)

MaliMaliPmMcuOff(MaliMaliPmmcuoffFtraceEvent)

MaliMaliPmMcuOn(MaliMaliPmmcuonFtraceEvent)

MaliMaliPmMcuOnCoreAttrUpdatePend(MaliMaliPmmcuoncoreattrupdatependFtraceEvent)

MaliMaliPmMcuOnGlbReinitPend(MaliMaliPmmcuonglbreinitpendFtraceEvent)

MaliMaliPmMcuOnHalt(MaliMaliPmmcuonhaltFtraceEvent)

MaliMaliPmMcuOnHwcntDisable(MaliMaliPmmcuonhwcntdisableFtraceEvent)

MaliMaliPmMcuOnHwcntEnable(MaliMaliPmmcuonhwcntenableFtraceEvent)

MaliMaliPmMcuOnPendHalt(MaliMaliPmmcuonpendhaltFtraceEvent)

MaliMaliPmMcuOnPendSleep(MaliMaliPmmcuonpendsleepFtraceEvent)

MaliMaliPmMcuOnSleepInitiate(MaliMaliPmmcuonsleepinitiateFtraceEvent)

MaliMaliPmMcuPendOff(MaliMaliPmmcupendoffFtraceEvent)

MaliMaliPmMcuPendOnReload(MaliMaliPmmcupendonreloadFtraceEvent)

MaliMaliPmMcuPowerDown(MaliMaliPmmcupowerdownFtraceEvent)

MaliMaliPmMcuResetWait(MaliMaliPmmcuresetwaitFtraceEvent)

Implementations

impl Event

pub fn encode(&self, buf: &mut impl BufMut)

Encodes the message to a buffer.

pub fn merge( field: &mut Option<Event>, tag: u32, wire_type: WireType, buf: &mut impl Buf, ctx: DecodeContext, ) -> Result<(), DecodeError>

Decodes an instance of the message from a buffer, and merges it into self.

pub fn encoded_len(&self) -> usize

Returns the encoded length of the message without a length delimiter.