micromegas::perfetto::protos

Module perf_events

Nested message and enum types in PerfEvents.

Modules

timebase

Nested message and enum types in Timebase.

Structs

RawEvent

Syscall-level description of the event, propagated to the perf_event_attr struct. Primarily for local use-cases, since the event availability and encoding is hardware-specific.

Timebase

What event to sample on, and how often. Commented from the perspective of its use in |PerfEventConfig|.

Tracepoint

Enums

Counter

Builtin counter names from the uapi header. Commented with their perf tool aliases. TODO(rsavitski): consider generating enums for cache events (should be finite), and generally make this list as extensive as possible. Excluding things like dynamic PMUs since those don’t fit into a static enum. Next id: 21

PerfClock

Subset of clocks that is supported by perf timestamping. CLOCK_TAI is excluded since it’s not expected to be used in practice, but would require additions to the trace clock synchronisation logic.