micromegas::telemetry_sink

Module reqwest

Modules

dns

DNS resolution

header

HTTP header types

redirect

Redirect Handling

retry

Retry requests

tls

TLS configuration and types

Structs

Body

An asynchronous request body.

Certificate

Represents a server X509 certificate.

Client

An asynchronous Client to make Requests with.

ClientBuilder

A ClientBuilder can be used to create a Client with custom configuration.

Error

The Errors that may occur when processing a Request.

Identity

Represents a private key and X509 cert as a client certificate.

Method

The Request Method (VERB)

NoProxy

A configuration for filtering out requests that shouldn’t be proxied

Proxy

Configuration of a proxy that a Client should pass requests to.

Request

A request which can be executed with Client::execute().

RequestBuilder

A builder to construct the properties of a Request.

Response

A Response to a submitted Request.

StatusCode

An HTTP status code (status-code in RFC 9110 et al.).

Upgraded

An upgraded HTTP connection.

Url

A parsed URL record.

Version

Represents a version of the HTTP spec.

Traits

IntoUrl

A trait to try to convert some type into a Url.

ResponseBuilderExt

Extension trait for http::response::Builder objects

Functions

get

Shortcut method to quickly make a GET request.

Type Aliases

Result

A Result alias where the Err case is reqwest::Error.