micromegas

Module tracing

low level tracing - has minimal dependencies

Modules

dispatch

Where events are recorded and eventually sent to a sink

errors

event

Structure to record events in memory

flush_monitor

FlushMonitor triggers the flush of the telemetry streams at regular interval.

guards

RAII-style guards

intern_string

Store dynamically-created strings in a global container. Strings are never released from the container.

levels

Verbosity management

logs

Events representing a process’s log

metrics

Events representing a measured scalar at a point in time

panic_hook

Reports panics as fatal log entries and shuts down the telemetry system

parsing

Manual parsing of dynamically sized events

prelude

process_info

Process metadata

property_set

Interned collection of PropertySet instances. Each PropertySet contains properties where the names and the values are statically allocated. The user is expected to manage the cardinality.

runtime

Runtime integration utilities for micromegas tracing

spans

Events reprensenting units of code execution

static_string_ref

StaticStringRef points to a string dependency keeping track of the codec. Necessary for unreal instrumentation where ansi and wide strings can coexist. In cases where the event format does not have to be compatible with unreal, StringId can be used.

string_id

StringId serializes the value of the pointer and the size of a UTF8 string. StaticStringRef should be prefered where wire compatibility with unreal is important.

time

System & monotonic tick count

Macros

async_span_scope

async_span_scope_named

debug

Logs a message at the debug level.

error

Logs a message at the error level.

fatal

Logs a message representing a crash or panic

fmetric

Records a float metric.

imetric

Records a integer metric.

info

Logs a message at the info level.

log

The standard logging macro.

log_enabled

Determines if a message logged at the specified level in that module will be logged.

span_scope

Records a sync span as two thread events

span_scope_named

Records a span with a name that is determined at runtime. The span name still needs to be statically allocated.

static_span_desc

trace

Logs a message at the trace level.

warn

Logs a message at the warn level.