micromegas::tracing

Module spans

Events reprensenting units of code execution

Structs

BeginAsyncNamedSpanEvent

BeginAsyncSpanEvent

BeginThreadNamedSpanEvent

BeginThreadSpanEvent

EndAsyncNamedSpanEvent

EndAsyncSpanEvent

EndThreadNamedSpanEvent

EndThreadSpanEvent

InstrumentedFuture

A wrapper that instruments a future with async span tracing

InstrumentedNamedFuture

A wrapper that instruments a future with named async span tracing

SpanLocation

SpanLocationRecord

SpanMetadata

SpanRecord

ThreadDepsQueue

ThreadEventQueue

Enums

ThreadDepsQueueAny

ThreadEventQueueAny

Traits

InstrumentFuture

Trait for adding instrumentation to futures

ThreadDepsQueueTypeIndex

ThreadEventQueueTypeIndex

Type Aliases

ThreadBlock

ThreadStream