micromegas::analytics::dfext

Module typed_column

Access to a RecordBatch’s columns

Functions

get_only_primitive_value

Retrieves the single primitive value from the first column of a single-record batch.

get_only_string_value

Retrieves the single string value from the first column of a single-record batch.

get_single_row_primitive_value

Retrieves the single primitive value from an indexed column of a single-record batch.

get_single_row_primitive_value_by_name

Retrieves the single primitive value from a named column of a single-record batch.

typed_column

Retrieves a typed column from a RecordBatch by index.

typed_column_by_name

Retrieves a typed column from a RecordBatch by name.