micromegas::analytics::lakehouse

Module jit_partitions

Management of process-specific partitions built on demand

Structs

JitPartitionConfig

Configuration for Just-In-Time (JIT) partition generation.

Functions

generate_process_jit_partitions

generate_process_jit_partitions lists the partitions that are needed to cover a time span for a specific process these partitions may not exist or they could be out of date Generates JIT partitions for a given time range filtered by process.

generate_process_jit_partitions_segment

Generates a segment of JIT partitions filtered by process.

generate_stream_jit_partitions

generate_stream_jit_partitions lists the partitiions that are needed to cover a time span these partitions may not exist or they could be out of date Generates JIT partitions for a given time range.

generate_stream_jit_partitions_segment

Generates a segment of JIT partitions.

is_jit_partition_up_to_date

is_jit_partition_up_to_date compares a partition spec with the partitions that exist to know if it should be recreated Checks if a JIT partition is up to date.

write_partition_from_blocks

Writes a partition from a set of blocks.