micromegas::analytics::lakehouse

Module partition_cache

In-memory copy of a subnet of the list of the partitions in the db

Structs

LivePartitionProvider

A QueryPartitionProvider that fetches partitions directly from the database.

NullPartitionProvider

A QueryPartitionProvider that always returns an empty list of partitions.

PartitionCache

PartitionCache allows to query partitions based on the insert_time range

PartitionWithMetadata

A partition with its file metadata loaded on-demand. This is used when you need both the partition data and its parquet metadata.

Traits

QueryPartitionProvider

A trait for providing queryable partitions.

Functions

partition_with_metadata

Convenience function to create a PartitionWithMetadata from an existing Partition. This loads the file metadata on-demand using the file_path.