micromegas::analytics::sql_arrow_bridge

Trait ColumnReader

pub trait ColumnReader {
    // Required methods
    fn extract_column_from_row(
        &self,
        row: &PgRow,
        struct_builder: &mut StructBuilder,
    ) -> Result<(), Error>;
    fn field(&self) -> Field;
}

A trait for reading a column from a database row and writing it to an Arrow StructBuilder.

Required Methods

fn extract_column_from_row( &self, row: &PgRow, struct_builder: &mut StructBuilder, ) -> Result<(), Error>

fn field(&self) -> Field