micromegas::client::perfetto_trace_client

Function format_perfetto_trace_stream

pub fn format_perfetto_trace_stream<'a>(
    client: &'a mut Client,
    process_id: &'a str,
    query_range: TimeRange,
    span_types: SpanTypes,
) -> impl Stream<Item = Result<Vec<u8>>> + 'a

Streams Perfetto trace chunks as they arrive from the server.

This function is useful for showing download progress to users since chunks are yielded as soon as they’re received rather than buffered.

Arguments

span_types - Types of spans to include: Thread, Async, or Both