micromegas::client::perfetto_trace_client

Function write_perfetto_trace

pub async fn write_perfetto_trace(
    client: &mut Client,
    process_id: &str,
    begin: DateTime<Utc>,
    end: DateTime<Utc>,
    out_filename: &str,
    span_types: SpanTypes,
) -> Result<()>

Writes a Perfetto trace to a file with configurable span types.

This function generates traces with thread spans, async spans, or both.

Arguments

span_types - Types of spans to include: Thread, Async, or Both