micromegas::perfetto::protos::chrome_compositor_state_machine

Struct MinorState

pub struct MinorState {

    pub commit_count: Option<i32>,
    pub current_frame_number: Option<i32>,
    pub last_frame_number_submit_performed: Option<i32>,
    pub last_frame_number_draw_performed: Option<i32>,
    pub last_frame_number_begin_main_frame_sent: Option<i32>,
    pub did_draw: Option<bool>,
    pub did_send_begin_main_frame_for_current_frame: Option<bool>,
    pub did_notify_begin_main_frame_not_expected_until: Option<bool>,
    pub did_notify_begin_main_frame_not_expected_soon: Option<bool>,
    pub wants_begin_main_frame_not_expected: Option<bool>,
    pub did_commit_during_frame: Option<bool>,
    pub did_invalidate_layer_tree_frame_sink: Option<bool>,
    pub did_perform_impl_side_invalidaion: Option<bool>,
    pub did_prepare_tiles: Option<bool>,
    pub consecutive_checkerboard_animations: Option<i32>,
    pub pending_submit_frames: Option<i32>,
    pub submit_frames_with_current_layer_tree_frame_sink: Option<i32>,
    pub needs_redraw: Option<bool>,
    pub needs_prepare_tiles: Option<bool>,
    pub needs_begin_main_frame: Option<bool>,
    pub needs_one_begin_impl_frame: Option<bool>,
    pub visible: Option<bool>,
    pub begin_frame_source_paused: Option<bool>,
    pub can_draw: Option<bool>,
    pub resourceless_draw: Option<bool>,
    pub has_pending_tree: Option<bool>,
    pub pending_tree_is_ready_for_activation: Option<bool>,
    pub active_tree_needs_first_draw: Option<bool>,
    pub active_tree_is_ready_to_draw: Option<bool>,
    pub did_create_and_initialize_first_layer_tree_frame_sink: Option<bool>,
    pub tree_priority: Option<i32>,
    pub scroll_handler_state: Option<i32>,
    pub critical_begin_main_frame_to_activate_is_fast: Option<bool>,
    pub main_thread_missed_last_deadline: Option<bool>,
    pub video_needs_begin_frames: Option<bool>,
    pub defer_begin_main_frame: Option<bool>,
    pub last_commit_had_no_updates: Option<bool>,
    pub did_draw_in_last_frame: Option<bool>,
    pub did_submit_in_last_frame: Option<bool>,
    pub needs_impl_side_invalidation: Option<bool>,
    pub current_pending_tree_is_impl_side: Option<bool>,
    pub previous_pending_tree_was_impl_side: Option<bool>,
    pub processing_animation_worklets_for_active_tree: Option<bool>,
    pub processing_animation_worklets_for_pending_tree: Option<bool>,
    pub processing_paint_worklets_for_pending_tree: Option<bool>,

}

Next id: 47

Fields
commit_count: Option<i32>current_frame_number: Option<i32>last_frame_number_submit_performed: Option<i32>last_frame_number_draw_performed: Option<i32>last_frame_number_begin_main_frame_sent: Option<i32>did_draw: Option<bool>did_send_begin_main_frame_for_current_frame: Option<bool>did_notify_begin_main_frame_not_expected_until: Option<bool>did_notify_begin_main_frame_not_expected_soon: Option<bool>wants_begin_main_frame_not_expected: Option<bool>did_commit_during_frame: Option<bool>did_invalidate_layer_tree_frame_sink: Option<bool>did_perform_impl_side_invalidaion: Option<bool>did_prepare_tiles: Option<bool>consecutive_checkerboard_animations: Option<i32>pending_submit_frames: Option<i32>submit_frames_with_current_layer_tree_frame_sink: Option<i32>needs_redraw: Option<bool>needs_prepare_tiles: Option<bool>needs_begin_main_frame: Option<bool>needs_one_begin_impl_frame: Option<bool>visible: Option<bool>begin_frame_source_paused: Option<bool>can_draw: Option<bool>resourceless_draw: Option<bool>has_pending_tree: Option<bool>pending_tree_is_ready_for_activation: Option<bool>active_tree_needs_first_draw: Option<bool>active_tree_is_ready_to_draw: Option<bool>did_create_and_initialize_first_layer_tree_frame_sink: Option<bool>tree_priority: Option<i32>scroll_handler_state: Option<i32>critical_begin_main_frame_to_activate_is_fast: Option<bool>main_thread_missed_last_deadline: Option<bool>video_needs_begin_frames: Option<bool>defer_begin_main_frame: Option<bool>last_commit_had_no_updates: Option<bool>did_draw_in_last_frame: Option<bool>did_submit_in_last_frame: Option<bool>needs_impl_side_invalidation: Option<bool>current_pending_tree_is_impl_side: Option<bool>previous_pending_tree_was_impl_side: Option<bool>processing_animation_worklets_for_active_tree: Option<bool>processing_animation_worklets_for_pending_tree: Option<bool>processing_paint_worklets_for_pending_tree: Option<bool>
Implementations

impl MinorState

pub fn commit_count(&self) -> i32

Returns the value of commit_count, or the default value if commit_count is unset.

pub fn current_frame_number(&self) -> i32

Returns the value of current_frame_number, or the default value if current_frame_number is unset.

pub fn last_frame_number_submit_performed(&self) -> i32

Returns the value of last_frame_number_submit_performed, or the default value if last_frame_number_submit_performed is unset.

pub fn last_frame_number_draw_performed(&self) -> i32

Returns the value of last_frame_number_draw_performed, or the default value if last_frame_number_draw_performed is unset.

pub fn last_frame_number_begin_main_frame_sent(&self) -> i32

Returns the value of last_frame_number_begin_main_frame_sent, or the default value if last_frame_number_begin_main_frame_sent is unset.

pub fn did_draw(&self) -> bool

Returns the value of did_draw, or the default value if did_draw is unset.

pub fn did_send_begin_main_frame_for_current_frame(&self) -> bool

Returns the value of did_send_begin_main_frame_for_current_frame, or the default value if did_send_begin_main_frame_for_current_frame is unset.

pub fn did_notify_begin_main_frame_not_expected_until(&self) -> bool

Returns the value of did_notify_begin_main_frame_not_expected_until, or the default value if did_notify_begin_main_frame_not_expected_until is unset.

pub fn did_notify_begin_main_frame_not_expected_soon(&self) -> bool

Returns the value of did_notify_begin_main_frame_not_expected_soon, or the default value if did_notify_begin_main_frame_not_expected_soon is unset.

pub fn wants_begin_main_frame_not_expected(&self) -> bool

Returns the value of wants_begin_main_frame_not_expected, or the default value if wants_begin_main_frame_not_expected is unset.

pub fn did_commit_during_frame(&self) -> bool

Returns the value of did_commit_during_frame, or the default value if did_commit_during_frame is unset.

pub fn did_invalidate_layer_tree_frame_sink(&self) -> bool

Returns the value of did_invalidate_layer_tree_frame_sink, or the default value if did_invalidate_layer_tree_frame_sink is unset.

pub fn did_perform_impl_side_invalidaion(&self) -> bool

Returns the value of did_perform_impl_side_invalidaion, or the default value if did_perform_impl_side_invalidaion is unset.

pub fn did_prepare_tiles(&self) -> bool

Returns the value of did_prepare_tiles, or the default value if did_prepare_tiles is unset.

pub fn consecutive_checkerboard_animations(&self) -> i32

Returns the value of consecutive_checkerboard_animations, or the default value if consecutive_checkerboard_animations is unset.

pub fn pending_submit_frames(&self) -> i32

Returns the value of pending_submit_frames, or the default value if pending_submit_frames is unset.

pub fn submit_frames_with_current_layer_tree_frame_sink(&self) -> i32

Returns the value of submit_frames_with_current_layer_tree_frame_sink, or the default value if submit_frames_with_current_layer_tree_frame_sink is unset.

pub fn needs_redraw(&self) -> bool

Returns the value of needs_redraw, or the default value if needs_redraw is unset.

pub fn needs_prepare_tiles(&self) -> bool

Returns the value of needs_prepare_tiles, or the default value if needs_prepare_tiles is unset.

pub fn needs_begin_main_frame(&self) -> bool

Returns the value of needs_begin_main_frame, or the default value if needs_begin_main_frame is unset.

pub fn needs_one_begin_impl_frame(&self) -> bool

Returns the value of needs_one_begin_impl_frame, or the default value if needs_one_begin_impl_frame is unset.

pub fn visible(&self) -> bool

Returns the value of visible, or the default value if visible is unset.

pub fn begin_frame_source_paused(&self) -> bool

Returns the value of begin_frame_source_paused, or the default value if begin_frame_source_paused is unset.

pub fn can_draw(&self) -> bool

Returns the value of can_draw, or the default value if can_draw is unset.

pub fn resourceless_draw(&self) -> bool

Returns the value of resourceless_draw, or the default value if resourceless_draw is unset.

pub fn has_pending_tree(&self) -> bool

Returns the value of has_pending_tree, or the default value if has_pending_tree is unset.

pub fn pending_tree_is_ready_for_activation(&self) -> bool

Returns the value of pending_tree_is_ready_for_activation, or the default value if pending_tree_is_ready_for_activation is unset.

pub fn active_tree_needs_first_draw(&self) -> bool

Returns the value of active_tree_needs_first_draw, or the default value if active_tree_needs_first_draw is unset.

pub fn active_tree_is_ready_to_draw(&self) -> bool

Returns the value of active_tree_is_ready_to_draw, or the default value if active_tree_is_ready_to_draw is unset.

pub fn did_create_and_initialize_first_layer_tree_frame_sink(&self) -> bool

Returns the value of did_create_and_initialize_first_layer_tree_frame_sink, or the default value if did_create_and_initialize_first_layer_tree_frame_sink is unset.

pub fn tree_priority(&self) -> TreePriority

Returns the enum value of tree_priority, or the default if the field is unset or set to an invalid enum value.

pub fn set_tree_priority(&mut self, value: TreePriority)

Sets tree_priority to the provided enum value.

pub fn scroll_handler_state(&self) -> ScrollHandlerState

Returns the enum value of scroll_handler_state, or the default if the field is unset or set to an invalid enum value.

pub fn set_scroll_handler_state(&mut self, value: ScrollHandlerState)

Sets scroll_handler_state to the provided enum value.

pub fn critical_begin_main_frame_to_activate_is_fast(&self) -> bool

Returns the value of critical_begin_main_frame_to_activate_is_fast, or the default value if critical_begin_main_frame_to_activate_is_fast is unset.

pub fn main_thread_missed_last_deadline(&self) -> bool

Returns the value of main_thread_missed_last_deadline, or the default value if main_thread_missed_last_deadline is unset.

pub fn video_needs_begin_frames(&self) -> bool

Returns the value of video_needs_begin_frames, or the default value if video_needs_begin_frames is unset.

pub fn defer_begin_main_frame(&self) -> bool

Returns the value of defer_begin_main_frame, or the default value if defer_begin_main_frame is unset.

pub fn last_commit_had_no_updates(&self) -> bool

Returns the value of last_commit_had_no_updates, or the default value if last_commit_had_no_updates is unset.

pub fn did_draw_in_last_frame(&self) -> bool

Returns the value of did_draw_in_last_frame, or the default value if did_draw_in_last_frame is unset.

pub fn did_submit_in_last_frame(&self) -> bool

Returns the value of did_submit_in_last_frame, or the default value if did_submit_in_last_frame is unset.

pub fn needs_impl_side_invalidation(&self) -> bool

Returns the value of needs_impl_side_invalidation, or the default value if needs_impl_side_invalidation is unset.

pub fn current_pending_tree_is_impl_side(&self) -> bool

Returns the value of current_pending_tree_is_impl_side, or the default value if current_pending_tree_is_impl_side is unset.

pub fn previous_pending_tree_was_impl_side(&self) -> bool

Returns the value of previous_pending_tree_was_impl_side, or the default value if previous_pending_tree_was_impl_side is unset.

pub fn processing_animation_worklets_for_active_tree(&self) -> bool

Returns the value of processing_animation_worklets_for_active_tree, or the default value if processing_animation_worklets_for_active_tree is unset.

pub fn processing_animation_worklets_for_pending_tree(&self) -> bool

Returns the value of processing_animation_worklets_for_pending_tree, or the default value if processing_animation_worklets_for_pending_tree is unset.

pub fn processing_paint_worklets_for_pending_tree(&self) -> bool

Returns the value of processing_paint_worklets_for_pending_tree, or the default value if processing_paint_worklets_for_pending_tree is unset.