micromegas::perfetto::protos

Enum VmstatCounters

#[repr(i32)]
pub enum VmstatCounters {

    VmstatUnspecified = 0,
    VmstatNrFreePages = 1,
    VmstatNrAllocBatch = 2,
    VmstatNrInactiveAnon = 3,
    VmstatNrActiveAnon = 4,
    VmstatNrInactiveFile = 5,
    VmstatNrActiveFile = 6,
    VmstatNrUnevictable = 7,
    VmstatNrMlock = 8,
    VmstatNrAnonPages = 9,
    VmstatNrMapped = 10,
    VmstatNrFilePages = 11,
    VmstatNrDirty = 12,
    VmstatNrWriteback = 13,
    VmstatNrSlabReclaimable = 14,
    VmstatNrSlabUnreclaimable = 15,
    VmstatNrPageTablePages = 16,
    VmstatNrKernelStack = 17,
    VmstatNrOverhead = 18,
    VmstatNrUnstable = 19,
    VmstatNrBounce = 20,
    VmstatNrVmscanWrite = 21,
    VmstatNrVmscanImmediateReclaim = 22,
    VmstatNrWritebackTemp = 23,
    VmstatNrIsolatedAnon = 24,
    VmstatNrIsolatedFile = 25,
    VmstatNrShmem = 26,
    VmstatNrDirtied = 27,
    VmstatNrWritten = 28,
    VmstatNrPagesScanned = 29,
    VmstatWorkingsetRefault = 30,
    VmstatWorkingsetActivate = 31,
    VmstatWorkingsetNodereclaim = 32,
    VmstatNrAnonTransparentHugepages = 33,
    VmstatNrFreeCma = 34,
    VmstatNrSwapcache = 35,
    VmstatNrDirtyThreshold = 36,
    VmstatNrDirtyBackgroundThreshold = 37,
    VmstatPgpgin = 38,
    VmstatPgpgout = 39,
    VmstatPgpgoutclean = 40,
    VmstatPswpin = 41,
    VmstatPswpout = 42,
    VmstatPgallocDma = 43,
    VmstatPgallocNormal = 44,
    VmstatPgallocMovable = 45,
    VmstatPgfree = 46,
    VmstatPgactivate = 47,
    VmstatPgdeactivate = 48,
    VmstatPgfault = 49,
    VmstatPgmajfault = 50,
    VmstatPgrefillDma = 51,
    VmstatPgrefillNormal = 52,
    VmstatPgrefillMovable = 53,
    VmstatPgstealKswapdDma = 54,
    VmstatPgstealKswapdNormal = 55,
    VmstatPgstealKswapdMovable = 56,
    VmstatPgstealDirectDma = 57,
    VmstatPgstealDirectNormal = 58,
    VmstatPgstealDirectMovable = 59,
    VmstatPgscanKswapdDma = 60,
    VmstatPgscanKswapdNormal = 61,
    VmstatPgscanKswapdMovable = 62,
    VmstatPgscanDirectDma = 63,
    VmstatPgscanDirectNormal = 64,
    VmstatPgscanDirectMovable = 65,
    VmstatPgscanDirectThrottle = 66,
    VmstatPginodesteal = 67,
    VmstatSlabsScanned = 68,
    VmstatKswapdInodesteal = 69,
    VmstatKswapdLowWmarkHitQuickly = 70,
    VmstatKswapdHighWmarkHitQuickly = 71,
    VmstatPageoutrun = 72,
    VmstatAllocstall = 73,
    VmstatPgrotated = 74,
    VmstatDropPagecache = 75,
    VmstatDropSlab = 76,
    VmstatPgmigrateSuccess = 77,
    VmstatPgmigrateFail = 78,
    VmstatCompactMigrateScanned = 79,
    VmstatCompactFreeScanned = 80,
    VmstatCompactIsolated = 81,
    VmstatCompactStall = 82,
    VmstatCompactFail = 83,
    VmstatCompactSuccess = 84,
    VmstatCompactDaemonWake = 85,
    VmstatUnevictablePgsCulled = 86,
    VmstatUnevictablePgsScanned = 87,
    VmstatUnevictablePgsRescued = 88,
    VmstatUnevictablePgsMlocked = 89,
    VmstatUnevictablePgsMunlocked = 90,
    VmstatUnevictablePgsCleared = 91,
    VmstatUnevictablePgsStranded = 92,
    VmstatNrZspages = 93,
    VmstatNrIonHeap = 94,
    VmstatNrGpuHeap = 95,
    VmstatAllocstallDma = 96,
    VmstatAllocstallMovable = 97,
    VmstatAllocstallNormal = 98,
    VmstatCompactDaemonFreeScanned = 99,
    VmstatCompactDaemonMigrateScanned = 100,
    VmstatNrFastrpc = 101,
    VmstatNrIndirectlyReclaimable = 102,
    VmstatNrIonHeapPool = 103,
    VmstatNrKernelMiscReclaimable = 104,
    VmstatNrShadowCallStackBytes = 105,
    VmstatNrShmemHugepages = 106,
    VmstatNrShmemPmdmapped = 107,
    VmstatNrUnreclaimablePages = 108,
    VmstatNrZoneActiveAnon = 109,
    VmstatNrZoneActiveFile = 110,
    VmstatNrZoneInactiveAnon = 111,
    VmstatNrZoneInactiveFile = 112,
    VmstatNrZoneUnevictable = 113,
    VmstatNrZoneWritePending = 114,
    VmstatOomKill = 115,
    VmstatPglazyfree = 116,
    VmstatPglazyfreed = 117,
    VmstatPgrefill = 118,
    VmstatPgscanDirect = 119,
    VmstatPgscanKswapd = 120,
    VmstatPgskipDma = 121,
    VmstatPgskipMovable = 122,
    VmstatPgskipNormal = 123,
    VmstatPgstealDirect = 124,
    VmstatPgstealKswapd = 125,
    VmstatSwapRa = 126,
    VmstatSwapRaHit = 127,
    VmstatWorkingsetRestore = 128,
    VmstatAllocstallDevice = 129,
    VmstatAllocstallDma32 = 130,
    VmstatBalloonDeflate = 131,
    VmstatBalloonInflate = 132,
    VmstatBalloonMigrate = 133,
    VmstatCmaAllocFail = 134,
    VmstatCmaAllocSuccess = 135,
    VmstatNrFileHugepages = 136,
    VmstatNrFilePmdmapped = 137,
    VmstatNrFollPinAcquired = 138,
    VmstatNrFollPinReleased = 139,
    VmstatNrSecPageTablePages = 140,
    VmstatNrShadowCallStack = 141,
    VmstatNrSwapcached = 142,
    VmstatNrThrottledWritten = 143,
    VmstatPgallocDevice = 144,
    VmstatPgallocDma32 = 145,
    VmstatPgdemoteDirect = 146,
    VmstatPgdemoteKswapd = 147,
    VmstatPgreuse = 148,
    VmstatPgscanAnon = 149,
    VmstatPgscanFile = 150,
    VmstatPgskipDevice = 151,
    VmstatPgskipDma32 = 152,
    VmstatPgstealAnon = 153,
    VmstatPgstealFile = 154,
    VmstatThpCollapseAlloc = 155,
    VmstatThpCollapseAllocFailed = 156,
    VmstatThpDeferredSplitPage = 157,
    VmstatThpFaultAlloc = 158,
    VmstatThpFaultFallback = 159,
    VmstatThpFaultFallbackCharge = 160,
    VmstatThpFileAlloc = 161,
    VmstatThpFileFallback = 162,
    VmstatThpFileFallbackCharge = 163,
    VmstatThpFileMapped = 164,
    VmstatThpMigrationFail = 165,
    VmstatThpMigrationSplit = 166,
    VmstatThpMigrationSuccess = 167,
    VmstatThpScanExceedNonePte = 168,
    VmstatThpScanExceedSharePte = 169,
    VmstatThpScanExceedSwapPte = 170,
    VmstatThpSplitPage = 171,
    VmstatThpSplitPageFailed = 172,
    VmstatThpSplitPmd = 173,
    VmstatThpSwpout = 174,
    VmstatThpSwpoutFallback = 175,
    VmstatThpZeroPageAlloc = 176,
    VmstatThpZeroPageAllocFailed = 177,
    VmstatVmaLockAbort = 178,
    VmstatVmaLockMiss = 179,
    VmstatVmaLockRetry = 180,
    VmstatVmaLockSuccess = 181,
    VmstatWorkingsetActivateAnon = 182,
    VmstatWorkingsetActivateFile = 183,
    VmstatWorkingsetNodes = 184,
    VmstatWorkingsetRefaultAnon = 185,
    VmstatWorkingsetRefaultFile = 186,
    VmstatWorkingsetRestoreAnon = 187,
    VmstatWorkingsetRestoreFile = 188,

}

Counter definitions for Linux’s /proc/vmstat.

Variants

VmstatUnspecified = 0

VmstatNrFreePages = 1

VmstatNrAllocBatch = 2

VmstatNrInactiveAnon = 3

VmstatNrActiveAnon = 4

VmstatNrInactiveFile = 5

VmstatNrActiveFile = 6

VmstatNrUnevictable = 7

VmstatNrMlock = 8

VmstatNrAnonPages = 9

VmstatNrMapped = 10

VmstatNrFilePages = 11

VmstatNrDirty = 12

VmstatNrWriteback = 13

VmstatNrSlabReclaimable = 14

VmstatNrSlabUnreclaimable = 15

VmstatNrPageTablePages = 16

VmstatNrKernelStack = 17

VmstatNrOverhead = 18

VmstatNrUnstable = 19

VmstatNrBounce = 20

VmstatNrVmscanWrite = 21

VmstatNrVmscanImmediateReclaim = 22

VmstatNrWritebackTemp = 23

VmstatNrIsolatedAnon = 24

VmstatNrIsolatedFile = 25

VmstatNrShmem = 26

VmstatNrDirtied = 27

VmstatNrWritten = 28

VmstatNrPagesScanned = 29

VmstatWorkingsetRefault = 30

VmstatWorkingsetActivate = 31

VmstatWorkingsetNodereclaim = 32

VmstatNrAnonTransparentHugepages = 33

VmstatNrFreeCma = 34

VmstatNrSwapcache = 35

VmstatNrDirtyThreshold = 36

VmstatNrDirtyBackgroundThreshold = 37

VmstatPgpgin = 38

VmstatPgpgout = 39

VmstatPgpgoutclean = 40

VmstatPswpin = 41

VmstatPswpout = 42

VmstatPgallocDma = 43

VmstatPgallocNormal = 44

VmstatPgallocMovable = 45

VmstatPgfree = 46

VmstatPgactivate = 47

VmstatPgdeactivate = 48

VmstatPgfault = 49

VmstatPgmajfault = 50

VmstatPgrefillDma = 51

VmstatPgrefillNormal = 52

VmstatPgrefillMovable = 53

VmstatPgstealKswapdDma = 54

VmstatPgstealKswapdNormal = 55

VmstatPgstealKswapdMovable = 56

VmstatPgstealDirectDma = 57

VmstatPgstealDirectNormal = 58

VmstatPgstealDirectMovable = 59

VmstatPgscanKswapdDma = 60

VmstatPgscanKswapdNormal = 61

VmstatPgscanKswapdMovable = 62

VmstatPgscanDirectDma = 63

VmstatPgscanDirectNormal = 64

VmstatPgscanDirectMovable = 65

VmstatPgscanDirectThrottle = 66

VmstatPginodesteal = 67

VmstatSlabsScanned = 68

VmstatKswapdInodesteal = 69

VmstatKswapdLowWmarkHitQuickly = 70

VmstatKswapdHighWmarkHitQuickly = 71

VmstatPageoutrun = 72

VmstatAllocstall = 73

VmstatPgrotated = 74

VmstatDropPagecache = 75

VmstatDropSlab = 76

VmstatPgmigrateSuccess = 77

VmstatPgmigrateFail = 78

VmstatCompactMigrateScanned = 79

VmstatCompactFreeScanned = 80

VmstatCompactIsolated = 81

VmstatCompactStall = 82

VmstatCompactFail = 83

VmstatCompactSuccess = 84

VmstatCompactDaemonWake = 85

VmstatUnevictablePgsCulled = 86

VmstatUnevictablePgsScanned = 87

VmstatUnevictablePgsRescued = 88

VmstatUnevictablePgsMlocked = 89

VmstatUnevictablePgsMunlocked = 90

VmstatUnevictablePgsCleared = 91

VmstatUnevictablePgsStranded = 92

VmstatNrZspages = 93

VmstatNrIonHeap = 94

VmstatNrGpuHeap = 95

VmstatAllocstallDma = 96

VmstatAllocstallMovable = 97

VmstatAllocstallNormal = 98

VmstatCompactDaemonFreeScanned = 99

VmstatCompactDaemonMigrateScanned = 100

VmstatNrFastrpc = 101

VmstatNrIndirectlyReclaimable = 102

VmstatNrIonHeapPool = 103

VmstatNrKernelMiscReclaimable = 104

VmstatNrShadowCallStackBytes = 105

VmstatNrShmemHugepages = 106

VmstatNrShmemPmdmapped = 107

VmstatNrUnreclaimablePages = 108

VmstatNrZoneActiveAnon = 109

VmstatNrZoneActiveFile = 110

VmstatNrZoneInactiveAnon = 111

VmstatNrZoneInactiveFile = 112

VmstatNrZoneUnevictable = 113

VmstatNrZoneWritePending = 114

VmstatOomKill = 115

VmstatPglazyfree = 116

VmstatPglazyfreed = 117

VmstatPgrefill = 118

VmstatPgscanDirect = 119

VmstatPgscanKswapd = 120

VmstatPgskipDma = 121

VmstatPgskipMovable = 122

VmstatPgskipNormal = 123

VmstatPgstealDirect = 124

VmstatPgstealKswapd = 125

VmstatSwapRa = 126

VmstatSwapRaHit = 127

VmstatWorkingsetRestore = 128

VmstatAllocstallDevice = 129

VmstatAllocstallDma32 = 130

VmstatBalloonDeflate = 131

VmstatBalloonInflate = 132

VmstatBalloonMigrate = 133

VmstatCmaAllocFail = 134

VmstatCmaAllocSuccess = 135

VmstatNrFileHugepages = 136

VmstatNrFilePmdmapped = 137

VmstatNrFollPinAcquired = 138

VmstatNrFollPinReleased = 139

VmstatNrSecPageTablePages = 140

VmstatNrShadowCallStack = 141

VmstatNrSwapcached = 142

VmstatNrThrottledWritten = 143

VmstatPgallocDevice = 144

VmstatPgallocDma32 = 145

VmstatPgdemoteDirect = 146

VmstatPgdemoteKswapd = 147

VmstatPgreuse = 148

VmstatPgscanAnon = 149

VmstatPgscanFile = 150

VmstatPgskipDevice = 151

VmstatPgskipDma32 = 152

VmstatPgstealAnon = 153

VmstatPgstealFile = 154

VmstatThpCollapseAlloc = 155

VmstatThpCollapseAllocFailed = 156

VmstatThpDeferredSplitPage = 157

VmstatThpFaultAlloc = 158

VmstatThpFaultFallback = 159

VmstatThpFaultFallbackCharge = 160

VmstatThpFileAlloc = 161

VmstatThpFileFallback = 162

VmstatThpFileFallbackCharge = 163

VmstatThpFileMapped = 164

VmstatThpMigrationFail = 165

VmstatThpMigrationSplit = 166

VmstatThpMigrationSuccess = 167

VmstatThpScanExceedNonePte = 168

VmstatThpScanExceedSharePte = 169

VmstatThpScanExceedSwapPte = 170

VmstatThpSplitPage = 171

VmstatThpSplitPageFailed = 172

VmstatThpSplitPmd = 173

VmstatThpSwpout = 174

VmstatThpSwpoutFallback = 175

VmstatThpZeroPageAlloc = 176

VmstatThpZeroPageAllocFailed = 177

VmstatVmaLockAbort = 178

VmstatVmaLockMiss = 179

VmstatVmaLockRetry = 180

VmstatVmaLockSuccess = 181

VmstatWorkingsetActivateAnon = 182

VmstatWorkingsetActivateFile = 183

VmstatWorkingsetNodes = 184

VmstatWorkingsetRefaultAnon = 185

VmstatWorkingsetRefaultFile = 186

VmstatWorkingsetRestoreAnon = 187

VmstatWorkingsetRestoreFile = 188

Implementations

impl VmstatCounters

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of VmstatCounters.

pub fn from_i32(value: i32) -> Option<VmstatCounters>

👎Deprecated: Use the TryFrom<i32> implementation instead

Converts an i32 to a VmstatCounters, or None if value is not a valid variant.

impl VmstatCounters

pub fn as_str_name(&self) -> &'static str

String value of the enum field names used in the ProtoBuf definition.

The values are not transformed in any way and thus are considered stable (if the ProtoBuf definition does not change) and safe for programmatic use.

pub fn from_str_name(value: &str) -> Option<VmstatCounters>

Creates an enum from field names used in the ProtoBuf definition.