micromegas

Module servers

Embedable ingestion, analytics and maintenance services. The user is expected to provide their own authentication.

Modules

axum_utils

axum-utils : observability middleware

cron_task

scheduled task for daemon

flight_sql_service_impl

minimal FlightSQL protocol implementation

http_gateway

http server that redirects queries to the analytics server translating the response into json

ingestion

routes for ingestion server based on axum

key_ring

named keys for authentication

log_uri_service

log uris of http requests

maintenance

implementation of maintenance daemon keeping the lakehouse updated

perfetto

web server for perfetto traces

sqlinfo

metadata about this implementation of FlightSQL

tonic_auth_interceptor

authentication for the gRPC stack