micromegas::servers

Module maintenance

implementation of maintenance daemon keeping the lakehouse updated

Structs

EveryDayTask

task running once a day to materialize older partitions

EveryHourTask

task running once an hour to materialize recent partitions

EveryMinuteTask

task running once a minute to materialize recent partitions

EverySecondTask

task running once a second to materialize newest partitions

Functions

daemon

Starts the maintenance daemon, which runs various scheduled tasks.

get_global_views_with_update_group

Retrieves a list of global views that have an associated update group.

materialize_all_views

Materializes all views within a given time range.

run_tasks_forever

Runs a collection of CronTasks indefinitely.