micromegas::servers::maintenance

Function daemon

pub async fn daemon(
    lakehouse: Arc<LakehouseContext>,
    views_to_update: Vec<Arc<dyn View>>,
) -> Result<()>

Starts the maintenance daemon, which runs various scheduled tasks.

This function initializes and spawns several CronTasks for daily, hourly, minute, and second-based maintenance operations, such as data materialization and cleanup.

Arguments

lakehouse - The lakehouse context with shared metadata cache.

views_to_update - A vector of views that need to be updated by the daemon.