micromegas::servers::tonic_auth_interceptor

Function check_auth

pub fn check_auth(
    req: Request<()>,
    keyring: &KeyRing,
) -> Result<Request<()>, Status>

Checks the authentication of a Tonic request using a KeyRing.

This function checks for a Bearer token in the Authorization header and validates it against the provided KeyRing.