micromegas::telemetry_sink::stream_info

Function make_stream_info

pub fn make_stream_info<Block>(stream: &EventStream<Block>) -> StreamInfo
where
    Block: TracingBlock,
    <Block as TracingBlock>::Queue: HeterogeneousQueue,
    <<Block as TracingBlock>::Queue as ExtractDeps>::DepsQueue: HeterogeneousQueue,

Creates a StreamInfo object from an EventStream.

This function extracts metadata about the stream, including its process and stream IDs, and the serialized metadata for its dependencies and objects.